+++ desktop/js/JeedomConnect.js
@@ -14,6 +14,70 @@
  * You should have received a copy of the GNU General Public License
  * along with Jeedom. If not, see <http://www.gnu.org/licenses/>.
  */
+
+$('.eqLogicThumbnailContainer').off('click', '.widgetDisplayCard').on('click', '.widgetDisplayCard', function () {
+
+  var eqId = $(this).attr('data-widget_id');
+  editWidgetModal(eqId, true, true, true, true);
+
+})
+
+async function editWidgetModal(widgetId, removeAction, exit, duplicate, checkEquipment) {
+
+  if (checkEquipment) {
+
+    var mesEquipments = await getUsedByEquipement(widgetId);
+    var inEquipments = mesEquipments.result.names;
+  }
+  else {
+    var inEquipments = undefined;
+  }
+
+  var widgetToEdit = allWidgetsDetail.find(w => w.id == widgetId);
+  getWidgetModal({ title: "Editer un widget", eqId: widgetId, widget: widgetToEdit, removeAction: removeAction, exit: exit, duplicate: duplicate, inEquipments: inEquipments }, function (result) {
+    refreshWidgetDetails();
+    if (!exit) refreshWidgetsContent();
+  });
+
+}
+
+async function getUsedByEquipement(eqId) {
+  const result = await $.post({
+    url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
+    data: {
+      action: 'getWidgetExistance',
+      id: eqId
+    },
+    cache: false,
+    dataType: 'json',
+    async: false,
+  });
+
+  if (result.state != 'ok') {
+    $('#div_alert').showAlert({
+      message: result.result,
+      level: 'danger'
+    });
+  }
+
+  return result;
+}
+
+
+
+
+/**************************************************** */
+/**************************************************** */
+/**************************************************** */
+/**************************************************** */
+/**************************************************** */
+/**************************************************** */
+/**************************************************** */
+
+
+
+
+
 var allWidgetsDetail;
 var roomList;
 var roomListOptions;
@@ -71,44 +135,6 @@ if (typeof allJeedomData === 'undefined') {
 }
 
 
-function sortWidgets() {
-  if (jcOrderBy === undefined) jcOrderBy = 'object';
-
-  allWidgetsDetail.sort(SortByName);
-  if (jcOrderBy == 'object') {
-    allWidgetsDetail.sort(SortByRoom);
-  } else if (jcOrderBy == 'type') {
-    allWidgetsDetail.sort(SortByType);
-  }
-}
-
-function SortByName(a, b) {
-  var aName = a.name?.toLowerCase();
-  var bName = b.name?.toLowerCase();
-  return ((aName < bName) ? -1 : ((aName > bName) ? 1 : 0));
-}
-
-function SortByType(a, b) {
-  var aType = a.type.toLowerCase();
-  var bType = b.type.toLowerCase();
-  return ((aType < bType) ? -1 : ((aType > bType) ? 1 : 0));
-}
-
-function SortByRoom(a, b) {
-
-  var aObj = ('room' in a) ? roomList.find(r => r.id == a.room) || undefined : undefined;
-  var bObj = ('room' in b) ? roomList.find(r => r.id == b.room) || undefined : undefined;
-
-  var aRoom = (aObj != undefined) ? aObj.name.toLowerCase() : 'AAAucun';
-  var bRoom = (bObj != undefined) ? bObj.name.toLowerCase() : 'AAAucun';
-
-  return ((aRoom < bRoom) ? -1 : ((aRoom > bRoom) ? 1 : 0));
-}
-
-$('.eqLogicAttr[data-l1key=name]').change(function () {
-  $('.eqNameQrCode').text($('.eqLogicAttr[data-l1key=name]').val());
-})
-
 $.post({
   url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
   data: {
@@ -129,761 +155,11 @@ $.post({
 });
 
 
-$('.eqLogicThumbnailContainer').off('click', '.widgetDisplayCard').on('click', '.widgetDisplayCard', function () {
+var widgetsCat = [];//used for widgets option
+var cmdCat = []; //used for cmdList
+var imgCat = []; //used for img list
+var moreInfos = []; //used for moreInfos
 
-  var eqId = $(this).attr('data-widget_id');
-  editWidgetModal(eqId, true, true, true, true);
-
-})
-
-async function getUsedByEquipement(eqId) {
-  const result = await $.post({
-    url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-    data: {
-      action: 'getWidgetExistance',
-      id: eqId
-    },
-    cache: false,
-    dataType: 'json',
-    async: false,
-  });
-
-  if (result.state != 'ok') {
-    $('#div_alert').showAlert({
-      message: result.result,
-      level: 'danger'
-    });
-  }
-
-  return result;
-}
-
-
-async function editWidgetModal(widgetId, removeAction, exit, duplicate, checkEquipment) {
-
-  if (checkEquipment) {
-
-    var mesEquipments = await getUsedByEquipement(widgetId);
-    var inEquipments = mesEquipments.result.names;
-  }
-  else {
-    var inEquipments = undefined;
-  }
-
-  var widgetToEdit = allWidgetsDetail.find(w => w.id == widgetId);
-  getWidgetModal({ title: "Editer un widget", eqId: widgetId, widget: widgetToEdit, removeAction: removeAction, exit: exit, duplicate: duplicate, inEquipments: inEquipments }, function (result) {
-    refreshWidgetDetails();
-    if (!exit) refreshWidgetsContent();
-  });
-
-}
-
-
-function getWidgetModal(_options, _callback) {
-  // console.log("getWidgetModal option recues : ", _options)
-  if (!isset(_options)) {
-    return;
-  }
-  $("#widgetModal").dialog('destroy').remove();
-  if ($("#widgetModal").length == 0) {
-    $('body').append('<div id="widgetModal"></div>');
-
-    $("#widgetModal").dialog({
-      title: _options.title,
-      closeText: '',
-      autoOpen: false,
-      modal: true,
-      width: 1250,
-      height: 0.8 * $(window).height()
-    });
-    jQuery.ajaxSetup({
-      async: false
-    });
-    $('#widgetModal').load('index.php?v=d&plugin=JeedomConnect&modal=assistant.widgetModal.JeedomConnect');
-    jQuery.ajaxSetup({
-      async: true
-    });
-  }
-  setWidgetModalData(_options);
-
-  if (_options.inEquipments !== undefined && _options.inEquipments != '') {
-
-    $('#widgetInclusion').html('<span id="widgetExistInEquipement" style="display:none">' + _options.inEquipments.join(', ') + '</span>');
-    $('#widgetInclusion').append('Utilisé dans :<br/><ul>');
-
-    $.each(_options.inEquipments, function (index, value) {
-      $("#widgetInclusion").append('<li>' + value + '</li>');
-    });
-    $("#widgetInclusion").append('</ul>');
-    $('#widgetInclusion').css('display', 'block');
-  }
-
-  if (_options.removeAction != true) {
-    $('.widgetMenu .removeWidget').hide();
-    $('.widgetMenu .hideWidget').addClass('roundedRight');
-  }
-
-  if ($('#widgetOptions').attr('widget-id') == undefined || $('#widgetOptions').attr('widget-id') == '' || !(_options.duplicate)) {
-    $('.widgetMenu .duplicateWidget').hide();
-    $('.widgetMenu .saveWidget').addClass('roundedLeft');
-  }
-  else {
-    $('.widgetMenu .duplicateWidget').show();
-    $('.widgetMenu .saveWidget').removeClass('roundedLeft');
-  }
-
-  if ($('#widgetOptions').attr('widget-id') != undefined && $('#widgetOptions').attr('widget-id') != '') {
-    $('.autoFillWidgetCmds').hide();
-  }
-
-  if (_options.exit == true) {
-    $('.widgetMenu .saveWidget').attr('exit-attr', 'true');
-  }
-
-  $("#widgetModal").dialog({ title: _options.title });
-  $('#widgetModal').dialog('open');
-
-};
-
-
-
-
-/********************************************** */
-/********************************************** */
-/********************************************** */
-
-
-$('.eqLogicAction[data-action=addWidget]').off('click').on('click', function () {
-  getWidgetModal({ title: "Configuration du widget", removeAction: false, exit: true });
-})
-
-$('.eqLogicAction[data-action=showError]').off('click').on('click', function () {
-
-  var hide = ($('#spanWidgetErreur').text() == 'Erreur') ? true : false;
-  if (hide) {
-    $('.widgetDisplayCard').not(".hasError").hide();
-  }
-  else {
-    $('.widgetDisplayCard').show();
-  }
-
-
-  var typeSelected = $('#widgetTypeSelect').val();
-
-  if (typeSelected != 'none') {
-    $('.widgetDisplayCard').not("[data-widget_type=" + typeSelected + "]").hide();
-  }
-
-
-  $('.eqLogicThumbnailContainer').packery();
-
-  if (hide) {
-    $('#spanWidgetErreur').text('Tous');
-    $('.eqLogicAction[data-action=showError]').css('color', 'grey');
-  }
-  else {
-    $('#spanWidgetErreur').text('Erreur');
-    $('.eqLogicAction[data-action=showError]').css('color', 'red');
-  }
-  updateWidgetCount()
-
-})
-
-$('.eqLogicAction[data-action=addWidgetBulk]').off('click').on('click', function () {
-  $("#widgetModal").dialog('destroy').remove();
-  $('body').append('<div id="widgetModal"></div>');
-  $('#widgetModal').dialog({
-    title: "{{Ajout de widgets en masse}}",
-    width: 0.95 * $(window).width(),
-    height: 0.8 * $(window).height(),
-    modal: true,
-    closeText: ''
-  });
-  $('#widgetModal').load('index.php?v=d&plugin=JeedomConnect&modal=assistant.widgetBulkModal.JeedomConnect').dialog('open');
-})
-
-
-$('.eqLogicAction[data-action=showCommunity]').off('click').on('click', function () {
-  // $('.pluginInfo').toggle("slide", { direction: "right" }, 1000);
-  getSimpleModal({
-    title: "Forum",
-    width: 0.5 * $(window).width(),
-    fields: [{
-      type: "string",
-      value: $('.txtInfoPlugin').html()
-    },
-    {
-      type: "string",
-      id: "infoPluginModal",
-      value: $('.infoPlugin').html()
-    }],
-    buttons: {
-      "Fermer": function () {
-        $('#simpleModalAlert').hide();
-        $(this).dialog("close");
-      },
-      "Copier": function () {
-        copyDivToClipboard('#infoPluginModal', true)
-      }
-    }
-  }, function (result) { });
-
-
-});
-
-$('.eqLogicAction[data-action=showSummary]').off('click').on('click', function () {
-  $('body').append('<div id="widgetSummaryModal"></div>');
-  $('#widgetSummaryModal').dialog({
-    title: "{{Synthèse globale des widgets}}",
-    autoOpen: false,
-    modal: true,
-    closeText: '',
-    width: 0.9 * $(window).width(),
-    height: 0.8 * $(window).height(),
-    closeOnEscape: false,
-    // open: function(event, ui) { $(".ui-dialog-titlebar-close").hide(); },
-    close: function (ev, ui) { check_before_closing(); }
-  });
-  $('#widgetSummaryModal').load('index.php?v=d&plugin=JeedomConnect&modal=assistant.widgetSummary.JeedomConnect').dialog('open');
-})
-
-$('.eqLogicAction[data-action=showEquipmentSummary]').off('click').on('click', function () {
-  $('body').append('<div id="equipmentSummaryModal"></div>');
-  $('#equipmentSummaryModal').dialog({
-    title: "{{Synthèse des équipements JC}}",
-    autoOpen: false,
-    modal: true,
-    closeText: '',
-    width: 0.9 * $(window).width(),
-    height: 0.8 * $(window).height(),
-    closeOnEscape: false,
-    // open: function(event, ui) { $(".ui-dialog-titlebar-close").hide(); },
-    close: function (ev, ui) { check_before_closing(); }
-  });
-  $('#equipmentSummaryModal').load('index.php?v=d&plugin=JeedomConnect&modal=assistant.equipmentSummary.JeedomConnect').dialog('open');
-})
-
-$('.eqLogicAction[data-action=showNotifAll]').off('click').on('click', function () {
-  $('body').append('<div id="notifAllModal"></div>');
-  $('#notifAllModal').dialog({
-    title: "{{Configuration de la notification de \"tous\" les équipements}}",
-    autoOpen: false,
-    modal: true,
-    closeText: '',
-    width: 0.7 * $(window).width(),
-    height: 0.8 * $(window).height(),
-    closeOnEscape: false
-  });
-  $('#notifAllModal').load('index.php?v=d&plugin=JeedomConnect&modal=assistant.notifAll.JeedomConnect').dialog('open');
-})
-
-// Start Generic Types
-// HACK Remove when gentype config in plugin is not needed anymore
-function gotoGenTypeConfig() {
-  $('#md_modal').dialog({ title: "{{Objets / Pièces}}" });
-  $('#md_modal').load('index.php?v=d&plugin=JeedomConnect&modal=gentype.objects').dialog('open');
-}
-
-$('.eqLogicAction[data-action=gotoGenTypeConfig]').off('click').on('click', function () {
-  gotoGenTypeConfig();
-})
-// End Generic Types
-
-$('.eqLogicAttr[data-l1key=configuration][data-l2key=apiKey]').on('change', function () {
-  var key = $('.eqLogicAttr[data-l1key=configuration][data-l2key=apiKey]').value();
-  $('#img_config').attr("src", 'plugins/JeedomConnect/data/qrcodes/' + key + '.png');
-
-});
-
-$("#assistant-btn").click(function () {
-  openAssistantWidgetModal($('.eqLogicAttr[data-l1key=id]').value());
-
-});
-
-$(".btnAssistant").click(function (event) {
-  var id = $(this).closest('.eqLogicDisplayCard').data('eqlogic_id');
-  openAssistantWidgetModal(id, event);
-});
-
-function openAssistantWidgetModal(id, event) {
-  if (event !== undefined) event.stopPropagation();
-  $('#md_modal').dialog({ title: "{{Configuration de l'équipement}}" });
-  $('#md_modal').load('index.php?v=d&plugin=JeedomConnect&modal=assistant.JeedomConnect&eqLogicId=' + id).dialog('open');
-}
-
-$("#notifConfig-btn").click(function () {
-  openAssistantNotificationModal($('.eqLogicAttr[data-l1key=id]').value());
-});
-
-$(".btnNotification").click(function (event) {
-  var id = $(this).closest('.eqLogicDisplayCard').data('eqlogic_id');
-  openAssistantNotificationModal(id, event);
-});
-
-function openAssistantNotificationModal(id, event) {
-  if (event !== undefined) event.stopPropagation();
-  $('#md_modal').dialog({ title: "{{Configuration des notifications}}" });
-  $('#md_modal').load('index.php?v=d&plugin=JeedomConnect&modal=notifs.JeedomConnect&eqLogicId=' + id).dialog('open');
-}
-
-$('.jeedomConnect').off('click', '#export-btn').on('click', '#export-btn', function () {
-
-  var apiKeyVal = $('.eqLogicAttr[data-l1key=configuration][data-l2key=apiKey]').value();
-
-  $.post({
-    url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-    data: {
-      action: 'generateFile',
-      type: 'exportEqConf',
-      apiKey: apiKeyVal
-    },
-    dataType: 'json',
-    success: function (data) {
-      if (data.state != 'ok') {
-        $('#div_alert').showAlert({
-          message: data.result,
-          level: 'danger'
-        });
-      }
-      else {
-        download(apiKeyVal + '.json', JSON.stringify(data.result), true);
-      }
-    }
-  });
-
-})
-
-
-$('.jeedomConnect').off('click', '#exportAll-btn').on('click', '#exportAll-btn', function () {
-  var apiKey = $('.eqLogicAttr[data-l1key=configuration][data-l2key=apiKey]').value();
-
-  $.post({
-    url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-    data: {
-      'action': 'getConfig',
-      'apiKey': apiKey,
-      all: true
-    },
-    dataType: 'json',
-    success: function (data) {
-      //console.log("roomList ajax received : ", data) ;
-      if (data.state != 'ok') {
-        $('#div_alert').showAlert({
-          message: data.result,
-          level: 'danger'
-        });
-      }
-      else {
-        download(apiKey + '_DEBUG.json', JSON.stringify(data.result), true);
-      }
-    }
-  });
-
-
-});
-
-$('.jeedomConnect').off('click', '#copy-btn').on('click', '#copy-btn', function () {
-  var from = $('.eqLogicAttr[data-l1key=configuration][data-l2key=apiKey]').text();
-
-  allJCEquipmentsWithoutCurrent = allJCEquipments.filter(function (obj) {
-    return obj.apiKey !== from;
-  }).map(item => {
-    return {
-      id: item.apiKey,
-      name: item.name
-    };
-  });
-  getSimpleModal({
-    title: "Recopier vers quel(s) appareil(s)",
-    fields: [
-      { title: "Choix", type: "checkboxes", choices: allJCEquipmentsWithoutCurrent },
-      { type: "line" },
-      { title: "Inclure les perso", type: "radios", choices: [{ id: 'yes', name: 'Oui' }, { id: 'no', name: 'Non', selected: 'checked' },] },
-      { type: 'description', text: 'Attention, lors de la recopie vous perdez la configuration présente actuelle sur les équipements "cibles"' }
-    ]
-  }, function (result) {
-    if (result.checkboxes.length === 0) {
-      throw "Il faut sélectionner au moins un appareil ! ";
-    }
-    $.post({
-      url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-      data: {
-        'action': 'copyConfig',
-        'from': from,
-        'to': result.checkboxes,
-        'withCustom': (result.radio == 'yes')
-      },
-      dataType: 'json',
-      success: function (data) {
-        if (data.state != 'ok') {
-          $('#div_alert').showAlert({
-            message: data.result,
-            level: 'danger'
-          });
-        }
-        else {
-          $('#div_alert').showAlert({
-            message: "C'est fait !",
-            level: 'success'
-          });
-        }
-      }
-    });
-  });
-});
-
-$("#import-btn").click(function () {
-  $("#import-input").click();
-});
-
-$("#import-input").change(function () {
-  var key = $('.eqLogicAttr[data-l1key=configuration][data-l2key=apiKey]').value();
-  if ($(this).prop('files').length > 0) {
-    file = $(this).prop('files')[0];
-    var reader = new FileReader();
-    reader.onload = (function (theFile) {
-      return function (e) {
-        config = e.target.result;
-        $.post({
-          url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-          data: {
-            action: 'uploadWidgets',
-            import: 'eqConfig',
-            data: config,
-            apiKey: key
-          },
-          dataType: 'json',
-          error: function (error) {
-            $('#div_alert').showAlert({ message: data.result, level: 'danger' });
-          },
-          success: function (data) {
-            var levelType = (data.state != 'ok') ? 'danger' : 'success';
-            $('#div_alert').showAlert({ message: data.result, level: levelType });
-
-          },
-
-        });
-      };
-    })(file);
-    reader.readAsText(file);
-    $(this).prop("value", "");
-  }
-});
-
-
-$(".btRegenerateApiKey").click(function () {
-  var warning = "<i source='md' name='alert-outline' style='color:#ff0000' class='mdi mdi-alert-outline'></i>";
-
-  getSimpleModal({
-    title: "Confirmation",
-    fields: [
-      {
-        type: "string",
-        value: "Avant de réaliser cette opération, assurez-vous que l'application n'est pas lancée sur votre appareil.<br><br>Voulez-vous vraiment regénérer la clé API de cet équipement ?"
-      }]
-  }, function (result) {
-    $.post({
-      url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-      data: {
-        'action': 'regenerateApiKey',
-        'eqId': $('.eqLogicAttr[data-l1key=id]').value(),
-        'apiKey': $('.eqLogicAttr[data-l1key=configuration][data-l2key=apiKey]').value()
-      },
-      cache: false,
-      dataType: 'json',
-      async: false,
-      success: function (data) {
-        if (data.state != 'ok') {
-          $('#div_alert').showAlert({
-            message: data.result,
-            level: 'danger'
-          });
-        }
-        else {
-          // console.log('btRegenerateApiKey data :', data);
-          var ApiKey = data.result.newapikey;
-          // console.log('new api key', ApiKey);
-          $('.eqLogicAttr[data-l1key=configuration][data-l2key=apiKey]').text(ApiKey)
-          $('#img_config').attr("src", 'plugins/JeedomConnect/data/qrcodes/' + ApiKey + '.png');
-        }
-      },
-      error: function (error) {
-        console.log("error while regenerating the api key", error)
-      }
-    });
-  });
-})
-
-$("#qrcode-regenerate").click(function () {
-  var key = $('.eqLogicAttr[data-l1key=configuration][data-l2key=apiKey]').value();
-  $.post({
-    url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-    data: {
-      'action': 'generateQRcode',
-      'id': $('.eqLogicAttr[data-l1key=id]').value()
-    },
-    success: function () {
-      $('#img_config').attr("src", 'plugins/JeedomConnect/data/qrcodes/' + key + '.png?' + new Date().getTime());
-    },
-    error: function (error) {
-      console.log("error while generating qr code ", error)
-    }
-  });
-});
-
-$("#removeDevice").click(function () {
-  $.post({
-    url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-    data: {
-      'action': 'removeDevice',
-      'id': $('.eqLogicAttr[data-l1key=id]').value()
-    },
-    success: function () {
-      $('.eqLogicAttr[data-l1key=configuration][data-l2key=deviceName]').html('');
-    },
-    error: function (error) {
-      console.log("error to remove device ", error);
-    }
-  });
-});
-
-
-$("#butCol").click(function () {
-  $("#hidCol").toggle("slow");
-  document.getElementById("listCol").classList.toggle('col-lg-12');
-  document.getElementById("listCol").classList.toggle('col-lg-10');
-});
-
-
-$('#in_searchWidget').off('keyup').keyup(function () {
-  var search = $(this).value()
-  var widgetFilter = $("#widgetTypeSelect option:selected").val();
-
-  if (search == '') {
-    if (widgetFilter == 'none') {
-      $('.widgetDisplayCard').show()
-    }
-    else {
-      $('.widgetDisplayCard').each(function () {
-        widgetType = $(this).attr('data-widget_type');
-        if (widgetFilter == widgetType) {
-          $(this).closest('.widgetDisplayCard').show()
-        }
-      })
-    }
-    $('.eqLogicThumbnailContainer').packery()
-    updateWidgetCount();
-    return;
-  }
-
-
-  $('.widgetDisplayCard').hide()
-  search = normTextLower(search)
-  var text
-  var widgetId
-
-  $('.widgetDisplayCard').each(function () {
-    text = normTextLower($(this).children('.name').text())
-    widgetId = normTextLower($(this).attr('data-widget_id'))
-    widgetType = $(this).attr('data-widget_type');
-    if (text.indexOf(search) >= 0 || widgetId.indexOf(search) >= 0) {
-      if (widgetFilter == 'none' || widgetFilter == widgetType) {
-        $(this).closest('.widgetDisplayCard').show()
-      }
-    }
-  })
-  $('.eqLogicThumbnailContainer').packery()
-  updateWidgetCount();
-})
-
-$('#bt_resetSearchWidget').on('click', function () {
-  $('#in_searchWidget').val('').keyup()
-})
-
-
-$("#commandtab").sortable({ axis: "y", cursor: "move", items: ".cmd", placeholder: "ui-state-highlight", tolerance: "intersect", forcePlaceholderSize: true });
-
-function addCmdToTable(_cmd) {
-  if (!isset(_cmd)) {
-    var _cmd = { configuration: {} };
-  }
-  if (!isset(_cmd.configuration)) {
-    _cmd.configuration = {};
-  }
-  var tr = '<tr class="cmd" data-cmd_id="' + init(_cmd.id) + '">';
-  tr += '<td style="min-width:50px;width:70px;">';
-  tr += '<span class="cmdAttr" data-l1key="id"></span>';
-  tr += '</td>';
-  tr += '<td style="min-width:300px;width:350px;">';
-  tr += '<div class="row">';
-  tr += '<div class="col-xs-7">';
-  tr += '<input class="cmdAttr form-control input-sm" data-l1key="name" placeholder="{{Nom de la commande}}">';
-  tr += '<select class="cmdAttr form-control input-sm" data-l1key="value" style="display : none;margin-top : 5px;" title="{{Commande information liée}}">';
-  tr += '<option value="">{{Aucune}}</option>';
-  tr += '</select>';
-  tr += '</div>';
-  tr += '<div class="col-xs-5">';
-  tr += '<a class="cmdAction btn btn-default btn-sm" data-l1key="chooseIcon"><i class="fas fa-flag"></i> {{Icône}}</a>';
-  tr += '<span class="cmdAttr" data-l1key="display" data-l2key="icon" style="margin-left : 10px;"></span>';
-  tr += '</div>';
-  tr += '</div>';
-  tr += '</td>';
-  tr += '<td>';
-  tr += '<span class="type" style="display:none;" type="' + init(_cmd.type) + '">' + jeedom.cmd.availableType() + '</span>';
-  tr += '<span class="subType" subType="' + init(_cmd.subType) + '"></span>';
-  tr += '</td>';
-  tr += '<td style="min-width:120px;width:140px;">';
-  tr += '<div><label class="checkbox-inline"><input type="checkbox" class="cmdAttr checkbox-inline" data-l1key="isVisible" checked/>{{Afficher}}</label> ';
-  tr += '<label class="checkbox-inline"><input type="checkbox" class="cmdAttr checkbox-inline" data-l1key="isHistorized" checked/>{{Historiser}}</label> ';
-  tr += '<label class="checkbox-inline"><input type="checkbox" class="cmdAttr" data-l1key="display" data-l2key="invertBinary"/>{{Inverser}}</label></div>';
-  tr += '</td>';
-  tr += '<td style="min-width:180px;">';
-  tr += '<input class="cmdAttr form-control input-sm" data-l1key="configuration" data-l2key="minValue" placeholder="{{Min.}}" title="{{Min.}}" style="width:30%;display:inline-block;"/> ';
-  tr += '<input class="cmdAttr form-control input-sm" data-l1key="configuration" data-l2key="maxValue" placeholder="{{Max.}}" title="{{Max.}}" style="width:30%;display:inline-block;"/> ';
-  tr += '<input class="cmdAttr form-control input-sm" data-l1key="unite" placeholder="{{Unité}}" title="{{Unité}}" style="width:30%;display:inline-block;"/>';
-  tr += '</td>';
-  tr += '<td >';
-  tr += '<input style="min-width:30px;width:50px;" class="cmdAttr" data-l1key="order" placeholder="{{Ordre affichage}}"/> ';
-  tr += '</td>';
-  tr += '<td>';
-  if (is_numeric(_cmd.id)) {
-    tr += '<a class="btn btn-default btn-xs cmdAction" data-action="configure"><i class="fas fa-cogs"></i></a> ';
-    tr += '<a class="btn btn-default btn-xs cmdAction" data-action="test"><i class="fas fa-rss"></i> Tester</a>';
-  }
-  tr += '<i class="fas fa-minus-circle pull-right cmdAction cursor" data-action="remove" style="display:none"></i>';
-  tr += '</td>';
-  tr += '</tr>';
-
-  $('#table_cmd tbody.cmd_' + _cmd.type).append(tr);
-  var tr = $('#table_cmd tbody.cmd_' + _cmd.type + ' tr').last();
-
-  jeedom.eqLogic.builSelectCmd({
-    id: $('.eqLogicAttr[data-l1key=id]').value(),
-    filter: { type: 'info' },
-    error: function (error) {
-      $('#div_alert').showAlert({ message: error.message, level: 'danger' });
-    },
-    success: function (result) {
-      tr.find('.cmdAttr[data-l1key=value]').append(result);
-      tr.setValues(_cmd, '.cmdAttr');
-      jeedom.cmd.changeType(tr, init(_cmd.subType));
-    }
-  });
-}
-
-
-/*
-function addCmdToTable(_cmd) {
-  if (!isset(_cmd)) {
-    var _cmd = {configuration: {}};
-  }
-  if (!isset(_cmd.configuration)) {
-    _cmd.configuration = {};
-  }
-  if (init(_cmd.type) == 'info') {
-    var disabled = (init(_cmd.configuration.virtualAction) == '1') ? 'disabled' : '';
-    var tr = '<tr class="cmd" data-cmd_id="' + init(_cmd.id) + '">';
-    tr += '<td>';
-    tr += '<span class="cmdAttr" data-l1key="id"></span>';
-    tr += '</td>';
-    tr += '<td>';
-    tr += '<input class="cmdAttr form-control input-sm" data-l1key="name" style="width : 140px;" placeholder="{{Nom de l\'info}}"></td>';
-    tr += '<td>';
-    tr += '<input class="cmdAttr form-control type input-sm" data-l1key="type" value="info" disabled style="margin-bottom : 5px;" />';
-    tr += '<input class="cmdAttr form-control type input-sm" data-l1key="subType" value="' + init(_cmd.subType) + '" disabled style="margin-bottom : 5px;" />';
-    tr += '</td><td>';
-    tr += '<span><label class="checkbox-inline"><input type="checkbox" class="cmdAttr checkbox-inline" data-l1key="isVisible" checked/>{{Afficher}}</label></span> ';
-    if (_cmd.subType == "binary") {
-        tr += '<span><label class="checkbox-inline"><input type="checkbox" class="cmdAttr checkbox-inline" data-l1key="isHistorized" checked/>{{Historiser}}</label></span> ';
-        tr += '<span><label class="checkbox-inline"><input type="checkbox" class="cmdAttr checkbox-inline" data-l1key="display" data-l2key="invertBinary" />{{Inverser}}</label></span>';
-    }
-    if (_cmd.subType == "numeric") {
-        tr += '<span><label class="checkbox-inline"><input type="checkbox" class="cmdAttr checkbox-inline" data-l1key="isHistorized" checked/>{{Historiser}}</label></span> ';
-        tr += '<input class="tooltips cmdAttr form-control input-sm" data-l1key="configuration" data-l2key="minValue" placeholder="{{Min}}" title="{{Min}}" style="width : 40%;display : inline-block;"> ';
-        tr += '<input class="tooltips cmdAttr form-control input-sm" data-l1key="configuration" data-l2key="maxValue" placeholder="{{Max}}" title="{{Max}}" style="width : 40%;display : inline-block;">';
-    }
-    tr += '</td>';
-    tr += '<td>';
-    if (is_numeric(_cmd.id)) {
-      tr += '<a class="btn btn-default btn-xs cmdAction" data-action="configure"><i class="fas fa-cogs"></i></a> ';
-      tr += '<a class="btn btn-default btn-xs cmdAction" data-action="test"><i class="fas fa-rss"></i> {{Tester}}</a>';
-    }
-    tr += '<i class="fas fa-minus-circle pull-right cmdAction cursor" data-action="remove"></i></td>';
-    tr += '</tr>';
-    $('#table_cmd tbody').append(tr);
-    $('#table_cmd tbody tr:last').setValues(_cmd, '.cmdAttr');
-    /*if (isset(_cmd.type)) {
-    $('#table_cmd tbody tr:last .cmdAttr[data-l1key=type]').value(init(_cmd.type));
-  }
-  jeedom.cmd.changeType($('#table_cmd tbody tr:last'), init(_cmd.subType));*/
-/*}
-
-if (init(_cmd.type) == 'action') {
-  var tr = '<tr class="cmd" data-cmd_id="' + init(_cmd.id) + '">';
-  tr += '<td>';
-  tr += '<span class="cmdAttr" data-l1key="id"></span>';
-  tr += '</td>';
-  tr += '<td>';
-  tr += '<div class="row">';
-  tr += '<div class="col-lg-6">';
-  tr += '<a class="cmdAction btn btn-default btn-sm" data-l1key="chooseIcon"><i class="fas fa-flag"></i> Icone</a>';
-  tr += '<span class="cmdAttr" data-l1key="display" data-l2key="icon" style="margin-left : 10px;"></span>';
-  tr += '</div>';
-  tr += '<div class="col-lg-6">';
-  tr += '<input class="cmdAttr form-control input-sm" data-l1key="name">';
-  tr += '</div>';
-  tr += '</div>';
-  tr += '<select class="cmdAttr form-control tooltips input-sm" data-l1key="value" style="width: 180px;display : none;margin-top : 5px;" title="{{La valeur de la commande vaut par défaut la commande}}">';
-  tr += '<option value="">Aucune</option>';
-  tr += '</select>';
-  tr += '</td>';
-  tr += '<td>';
-  tr += '<input class="cmdAttr form-control type input-sm" data-l1key="type" value="action" disabled style="margin-bottom : 5px;" />';
-  tr += '<input class="cmdAttr form-control type input-sm" data-l1key="subType" value="' + init(_cmd.subType) + '" disabled style="margin-bottom : 5px;" />';
-  tr += '</td>';
-  tr += '<td>';
-  tr += '<span><label class="checkbox-inline"><input type="checkbox" class="cmdAttr checkbox-inline" data-l1key="isVisible" checked/>{{Afficher}}</label></span> ';
-  tr += '</td>';
-  tr += '<td>';
-  if (is_numeric(_cmd.id)) {
-    tr += '<a class="btn btn-default btn-xs cmdAction" data-action="configure"><i class="fas fa-cogs"></i></a> ';
-    tr += '<a class="btn btn-default btn-xs cmdAction" data-action="test"><i class="fas fa-rss"></i> {{Tester}}</a>';
-  }
-  tr += '<i class="fas fa-minus-circle pull-right cmdAction cursor" data-action="remove"></i></td>';
-  tr += '</tr>';
-
-  $('#table_cmd tbody').append(tr);
-  //$('#table_cmd tbody tr:last').setValues(_cmd, '.cmdAttr');
-  var tr = $('#table_cmd tbody tr:last');
-  jeedom.eqLogic.builSelectCmd({
-    id: $('.eqLogicAttr[data-l1key=id]').value(),
-    filter: {type: 'info'},
-    error: function (error) {
-      $('#div_alert').showAlert({message: error.message, level: 'danger'});
-    },
-    success: function (result) {
-      tr.find('.cmdAttr[data-l1key=value]').append(result);
-      tr.setValues(_cmd, '.cmdAttr');
-      jeedom.cmd.changeType(tr, init(_cmd.subType));
-    }
-  });
-
-}
-}
-*/
-
-
-
-////////////////////////////////////////////////////////////
-///////////////  ACTION SUR FORM WIDGET ////////////////////
-////////////////////////////////////////////////////////////
-
-//used for widgets option
-var widgetsCat = [];
-//used for cmdList
-var cmdCat = [];
-//used for img list
-var imgCat = [];
 var widgetsList = (function () {
   var json = null;
   $.ajax({
@@ -911,9 +187,6 @@ var widgetsList = (function () {
   return json;
 })();
 
-//used for moreInfos
-var moreInfos = [];
-
 items = [];
 widgetsList.widgets.forEach(item => {
   items.push('<option value="' + item.type + '">' + item.name + '</option>');
@@ -922,449 +195,7 @@ $("#widgetsList-select").html(items.join(""));
 $("#room-input").html(roomListOptions);
 
 
-function setWidgetModalData(options) {
-  $('#summaryModal').dialog('destroy').remove();
 
-  refreshAddWidgets();
-
-  if (options.widget !== undefined) {
-    $('#widgetsList-select option[value="' + options.widget.type + '"]').prop('selected', true);
-    refreshAddWidgets();
-    //Enable
-    var enable = options.widget.enable ? "checked" : "";
-    $("#enable-input").prop('checked', enable);
-    var blockDetail = options.widget.blockDetail ? "checked" : "";
-    $("#blockDetail-input").prop('checked', blockDetail);
-
-    //Room
-    if (options.widget.room !== undefined) {
-      $('#room-input option[value="' + options.widget.room + '"]').prop('selected', true);
-    }
-
-    moreInfos = options.widget.moreInfos || [];
-    refreshMoreInfos();
-
-    $("#widgetOptions").attr('widget-id', options.eqId ?? '');
-
-    var widgetConfig = widgetsList.widgets.find(i => i.type == options.widget.type);
-    //  console.log("widgetsList => ", widgetsList) ;
-    //  console.log("widgetConfig => ", widgetConfig) ;
-    widgetConfig.options.forEach(option => {
-      if (option.category == "string" & options.widget[option.id] !== undefined) {
-        $("#" + option.id + "-input").val(options.widget[option.id]);
-      } else if (option.category == "binary" & options.widget[option.id] !== undefined) {
-        $("#" + option.id + "-input").prop('checked', options.widget[option.id] ? "checked" : "");
-      } else if (option.category == "color" & options.widget[option.id] !== undefined) {
-        $("#" + option.id + "-input").val(options.widget[option.id]);
-        $("#" + option.id + "-picker").val(options.widget[option.id]);
-      } else if (option.category == "cmd" & options.widget[option.id] !== undefined) {
-        $("#" + option.id + "-input").attr('cmdId', options.widget[option.id].id);
-        cmdHumanName = getHumanName(options.widget[option.id].id)
-        if (cmdHumanName != '') {
-          $("#" + option.id + "-input").val(cmdHumanName);
-          $("#" + option.id + "-input").attr('title', cmdHumanName);
-          refreshImgListOption();
-          refreshInfoSelect();
-        }
-        else {
-          $("#" + option.id + "-input").val('#' + options.widget[option.id].id + '#');
-        }
-        $("#" + option.id + "-input").attr('cmdType', options.widget[option.id].type);
-        $("#" + option.id + "-input").attr('cmdSubType', options.widget[option.id].subType);
-        if (options.widget[option.id].type == 'action') {
-          $("#confirm-div-" + option.id).css('display', '');
-          $("#secure-div-" + option.id).css('display', '');
-          $("#pwd-div-" + option.id).css('display', '');
-          $("#confirm-" + option.id).prop('checked', options.widget[option.id].confirm ? "checked" : "");
-          $("#secure-" + option.id).prop('checked', options.widget[option.id].secure ? "checked" : "");
-          $("#pwd-" + option.id).prop('checked', options.widget[option.id].pwd ? "checked" : "");
-        } else {
-          $("#confirm-div-" + option.id).css('display', 'none');
-          $("#secure-div-" + option.id).css('display', 'none');
-          $("#pwd-div-" + option.id).css('display', 'none');
-        }
-        if (options.widget[option.id].subType == 'slider' | options.widget[option.id].subType == 'numeric') {
-          $("#" + option.id + "-minInput").css('display', '');
-          $("#" + option.id + "-maxInput").css('display', '');
-          $("#" + option.id + "-minInput").val(options.widget[option.id].minValue);
-          $("#" + option.id + "-maxInput").val(options.widget[option.id].maxValue);
-        } else {
-          $("#" + option.id + "-minInput").css('display', 'none');
-          $("#" + option.id + "-maxInput").css('display', 'none');
-        }
-        if (options.widget[option.id].subType == 'binary' | options.widget[option.id].subType == 'numeric') {
-          $("#invert-div-" + option.id).css('display', '');
-          $("#invert-" + option.id).prop('checked', options.widget[option.id].invert ? "checked" : "");
-        } else {
-          $("#invert-div-" + option.id).css('display', 'none');
-        }
-        if (options.widget[option.id].subType == 'numeric') {
-          $("#" + option.id + "-unitInput").css('display', '');
-          $("#" + option.id + "-unitInput").val(options.widget[option.id].unit);
-        } else {
-          $("#" + option.id + "-unitInput").css('display', 'none');
-        }
-        if (options.widget[option.id].subType == 'slider') {
-          $("#" + option.id + "-stepInput").css('display', '');
-          $("#" + option.id + "-stepInput").val(options.widget[option.id].step);
-        }
-        else {
-          $("#" + option.id + "-stepInput").css('display', 'none');
-        }
-
-      } else if (option.category == "scenario" & options.widget[option.id] !== undefined) {
-        getScenarioHumanName({
-          id: options.widget[option.id],
-          error: function (error) { },
-          success: function (data) {
-            data.forEach(sc => {
-              if (sc['id'] == options.widget[option.id]) {
-                $("#" + option.id + "-input").attr('scId', options.widget[option.id]);
-                $("#" + option.id + "-input").val(sc['humanName']);
-              }
-            })
-          }
-        });
-        $('#optionScenario').css('display', 'block');
-
-        if (options.widget['options'] !== undefined &&
-          options.widget['options']['tags'] !== undefined &&
-          options.widget['options']['tags'] != '') {
-          getHumanNameFromCmdId({ alert: '#widget-alert', cmdIdData: options.widget['options']['tags'] }, function (result, _params) {
-            $('#tags-scenario-input').val(result);
-          });
-        }
-        if (options.widget['options'] !== undefined) {
-          $("#confirm-" + option.id).prop('checked', options.widget['options'].confirm ? "checked" : "");
-          $("#secure-" + option.id).prop('checked', options.widget['options'].secure ? "checked" : "");
-          $("#pwd-" + option.id).prop('checked', options.widget['options'].pwd ? "checked" : "");
-        }
-
-      } else if (option.category == "stringList" & options.widget[option.id] !== undefined) {
-        var selectedChoice = option.choices.find(s => s.id == options.widget[option.id]);
-        if (selectedChoice !== undefined) {
-          $('#' + option.id + '-input option[value="' + options.widget[option.id] + '"]').prop('selected', true);
-          if (option.id == "subtitle") {
-            $("#subtitle-input-value").val(selectedChoice.id)
-          }
-        } else if (option.id == "subtitle" & options.widget.subtitle !== undefined) {
-          $('#subtitle-input option[value="custom"]').prop('selected', true);
-          $("#subtitle-input-value").val(options.widget.subtitle)
-          $("#subtitle-input-value").css('display', 'block');
-          $("#subtitle-select").show();
-        }
-      } else if (option.category == "widgets" & options.widget[option.id] !== undefined) {
-        widgetsCat = options.widget[option.id];
-        refreshWidgetOption();
-      } else if (option.category == "cmdList" & options.widget[option.id] !== undefined) {
-        cmdCat = options.widget[option.id];
-        refreshCmdListOption(JSON.stringify(option.options));
-      } else if (option.category == "ifImgs" & options.widget[option.id] !== undefined) {
-        imgCat = options.widget[option.id];
-        refreshImgListOption();
-      } else if (option.category == "choicesList") {
-        option.choices.forEach(v => {
-          $("#" + v.id + "-jc-checkbox").prop('checked', options.widget[v.id] ? "checked" : "");
-        });
-      } else if (option.category == "img" & options.widget[option.id] !== undefined) {
-        $("#icon-div-" + option.id).html(iconToHtml(options.widget[option.id]));
-      }
-    });
-  }
-  refreshStrings();
-  loadSortable('all');
-}
-
-function refreshAddWidgets() {
-  widgetsCat = [];
-  cmdCat = [];
-  imgCat = [];
-  moreInfos = [];
-  var type = $("#widgetsList-select").val();
-  var widget = widgetsList.widgets.find(i => i.type == type);
-  showAutoFillWidgetCmds();
-
-  (type == 'jc') ? $('.searchForJCeq').show() : $('.searchForJCeq').hide();
-
-  $("#widgetImg").attr("src", "plugins/JeedomConnect/data/img/" + widget.img);
-
-  $("#widgetDescription").html(widget.description);
-
-  if (widget.variables) {
-    let varDescr = `Variables disponibles : <ul style="padding-left: 15px;">`;
-    widget.variables.forEach(v => {
-      varDescr += `<li>#${v.name}# : ${v.descr}</li>`;
-    });
-    varDescr += '</ul>';
-    $("#widgetVariables").html(varDescr);
-    $("#widgetVariables").show();
-  } else {
-    $("#widgetVariables").hide();
-  }
-
-  var items = [];
-
-  //Enable
-  option = `<li><div class='form-group'>
-  	<label class='col-xs-3 '>Actif</label>
-  	<div class='col-xs-9'><div class='input-group'><input type="checkbox" style="width:150px;" id="enable-input" checked></div></div></div></li>`;
-  items.push(option);
-
-  //Room
-  option = `<li><div class='form-group'>
-    <label class='col-xs-3 ${type == 'room' ? 'required' : ''}'>Pièce</label>
-    <div class='col-xs-9'><div class='input-group'><select style="width:340px;" id="room-input" value=''>
-    <option value="none">Sélection  d'une pièce</option>`;
-  option += roomListOptions;
-
-  if (type == 'room') {
-    option += `<option value="global">Global</option>`;
-  }
-  option += `</select></div></div></div></li>`;
-  items.push(option);
-
-  widget.options.forEach(option => {
-    var required = (option.required) ? "required" : "";
-    var description = (option.description == undefined) ? '' : option.description;
-    var curOption = `<li><div class='form-group' id="${option.id}-div">
-    <label class='col-xs-3  ${required}'   id="${option.id}-label">${option.name}</label>
-    <div class='col-xs-9' id="${option.id}-div-right">
-    <div class="description">${description}</div>`;
-
-    if (option.category == "cmd") {
-      isDisabled = isJcExpert ? '' : 'disabled';
-      curOption += `<table><tr class="cmd">
-            <td>
-              <input class='input-sm form-control roundedLeft needRefresh' style="width:250px;" id="${option.id}-input" value='' cmdId='' cmdType='' cmdSubType='' ${isDisabled} configtype='${option.type}' configsubtype='${option.subtype}' configlink='${option.value}'>
-              <td>
-                 <a class='btn btn-default btn-sm cursor bt_selectTrigger' tooltip='Choisir une commande' onclick="selectCmd('${option.id}', '${option.type}', '${option.subtype}', '${option.value}');">
-                    <i class='fas fa-list-alt'></i></a>
-                    </td>
-            </td>
-            <td>
-                <i class="mdi mdi-minus-circle" id="${option.id}-remove"
-                      style="color:rgb(185, 58, 62);font-size:16px;margin-right:10px;display:${option.required ? 'none' : ''};" aria-hidden="true" onclick="removeCmd('${option.id}')"></i>
-            </td>
-            <td>
-                    <div style="width:50px;margin-left:5px; display:none;" id="invert-div-${option.id}">
-                    <i class='fa fa-sync' title="Inverser"></i><input type="checkbox" style="margin-left:5px;" id="invert-${option.id}"></div>
-            </td>
-            <td>
-                  <div style="width:50px;margin-left:5px; display:none;" id="confirm-div-${option.id}">
-                  <i class='fa fa-question' title="Demander confirmation"></i><input type="checkbox" style="margin-left:5px;" id="confirm-${option.id}"></div>
-            </td>
-            <td>
-                    <div style="width:50px; display:none;" id="secure-div-${option.id}">
-                    <i class='fa fa-fingerprint' title="Sécuriser avec empreinte digitale"></i><input type="checkbox" style="margin-left:5px;" id="secure-${option.id}"  ></div>
-            </td>
-            <td>
-                    <div style="width:50px; display:none;" id="pwd-div-${option.id}">
-                    <i class='mdi mdi-numeric' title="Sécuriser avec un code"></i><input type="checkbox" style="margin-left:5px;" id="pwd-${option.id}"  ></div>
-            </td>
-            <td>
-                <input type="number" style="width:50px; display:none;" id="${option.id}-minInput" value='' placeholder="Min">
-            </td>
-            <td>
-                <input type="number" style="width:50px;margin-left:5px; display:none;" id="${option.id}-maxInput" value='' placeholder="Max">
-            </td>
-            <td>
-                <input type="number" step="0.1" style="width:50px;margin-left:5px; display:none;" id="${option.id}-stepInput" value='1' placeholder="Step">
-            </td>
-            <td>
-                <input style="width:50px; margin-left:5px; display:none;" id="${option.id}-unitInput" value='' placeholder="Unité">
-            </td></tr></table>
-                    `;
-
-      curOption += "</div></div></li>";
-
-    } else if (option.category == "string") {
-      type = (option.subtype != undefined) ? option.subtype : 'text';
-      if (option.subtype == "multiline") {
-        curOption += `<div class='input-group'>
-        <div style="display:flex"><textarea style="width:340px;" id="${option.id}-input" value=''></textarea>`;
-      } else {
-        curOption += `<div class='input-group'>
-        <div style="display:flex"><input type="${type}" style="width:340px;" id="${option.id}-input" value=''>`;
-      }
-
-      if (option.id == 'name') {
-        curOption += `
-              <div class="dropdown" id="name-select">
-              <a class="btn btn-default dropdown-toggle" data-toggle="dropdown" style="height" >
-              <i class="fas fa-plus-square"></i> </a>
-              <ul class="dropdown-menu infos-select" input="${option.id}-input">`;
-        if (widget.variables) {
-          widget.variables.forEach(v => {
-            curOption += `<li info="${v.name}" onclick="infoSelected('#${v.name}#', this)"><a href="#">#${v.name}#</a></li>`;
-          });
-        }
-        curOption += `</ul></div></div>`
-      }
-      curOption += `</div></div></div></li>`;
-    } else if (option.category == "binary") {
-      curOption += `<div class='input-group'><input type="checkbox" style="width:150px;" id="${option.id}-input"></div>
-         </div></div></li>`;
-    } else if (option.category == "color") {
-      curOption += `<div class='input-group'><input style="width:200px;" id="${option.id}-input" >
-      <input type="color" id="${option.id}-picker"  onchange="colorDefined(this, '${option.id}')">
-      </div>
-         </div></div></li>`;
-    } else if (option.category == "stringList") {
-      curOption += `<div class='input-group'><select style="width:340px;" id="${option.id}-input" onchange="subtitleSelected();">`;
-      if (!required) {
-        curOption += `<option value="none">Aucun</option>`;
-      }
-      option.choices.forEach(item => {
-        curOption += `<option value="${item.id}">${item.name}</option>`;
-      })
-      if (option.id == "subtitle") {
-        curOption += `<option value="custom">Personnalisé</option></select>`;
-        curOption += `<div style="display:flex">
-  					<textarea style="width:340px; margin-top:5px; display:none;" id="subtitle-input-value" value='none'></textarea>`;
-
-        curOption += `
-            <div class="dropdown" id="subtitle-select" style=" display:none;">
-            <a class="btn btn-default dropdown-toggle" data-toggle="dropdown" style="height" >
-            <i class="fas fa-plus-square"></i> </a>
-            <ul class="dropdown-menu infos-select" input="subtitle-input-value">`;
-        if (widget.variables) {
-          widget.variables.forEach(v => {
-            curOption += `<li info="${v.name}" onclick="infoSelected('#${v.name}#', this)"><a href="#">#${v.name}#</a></li>`;
-          });
-        }
-        curOption += `</ul></div></div>`
-      } else {
-        curOption += '</select>';
-      }
-      curOption += `</div></div></div></li>`;
-    } else if (option.category == "img") {
-      curOption += `
-              <a class="btn btn-success roundedRight" onclick="imagePicker(this)"><i class="fas fa-check-square">
-              </i> Choisir </a>
-              <a data-id="icon-div-${option.id}" id="icon-div-${option.id}" onclick="removeImage(this)"></a>
-              </div></div></li>`;
-
-
-    } else if (option.category == "widgets") {
-      var widgetChoices = [];
-      widgetsList.widgets.forEach(item => {
-        if (option.whiteList !== undefined) {
-          if (option.whiteList.includes(item.type)) {
-            widgetChoices.push(item.type);
-          }
-        } else if (option.blackList !== undefined) {
-          if (!option.blackList.includes(item.type)) {
-            widgetChoices.push(item.type);
-          }
-        } else {
-          widgetChoices.push(item.type);
-        }
-      })
-      curOption += `<span class="input-group-btn">
-              <a class="btn btn-default roundedRight" onclick="addWidgetOption('${widgetChoices.join(".")}')"><i class="fas fa-plus-square">
-              </i> Ajouter</a></span><div id="widget-option"></div>`;
-      curOption += `</div></div></li>`;
-    } else if (option.category == "cmdList") {
-      curOption += `<span class="input-group-btn">
-              <a class="btn btn-default roundedRight" onclick="addCmdOption('${JSON.stringify(option.options).replace(/"/g, '&quot;')}')"><i class="fas fa-plus-square">
-              </i> Ajouter</a></span><div id="cmdList-option" data-cmd-options="${JSON.stringify(option.options).replace(/"/g, '&quot;')}" style='margin-left:-150px;'></div>`;
-      curOption += `</div></div></li>`;
-    } else if (option.category == "ifImgs") {
-      curOption += `<span class="input-group-btn">
-              <a class="btn btn-default roundedRight" onclick="addImgOption('widget')"><i class="fas fa-plus-square">
-              </i> Ajouter</a></span><div id="imgList-option"></div>`;
-      curOption += `</div></div></li>`;
-    } else if (option.category == "scenario") {
-      curOption += `<div class='input-group'><input class='input-sm form-control roundedLeft' id="${option.id}-input" value='' scId='' disabled>
-    <span class='input-group-btn'><a class='btn btn-default btn-sm cursor bt_selectTrigger' tooltip='Choisir un scenario' onclick="selectScenario('${option.id}');">
-    <i class='fas fa-list-alt'></i></a></span></div>
-      <div id='optionScenario' style='display:none;'>
-        <div class="input-group input-group-sm" style="width: 100%">
-            <span class="input-group-addon roundedLeft" style="width: 100px">Tags</span>
-            <input style="width:100%;" class='input-sm form-control roundedRight title' type="string" id="tags-scenario-input" value="" placeholder="Si nécessaire indiquez des tags" />
-        </div>
-        <div class="" style="width: 100%;display: flex;">
-          <div class="input-group input-group-sm">
-            <span class="input-group-addon roundedLeft" style="width: 100px">Sécurité</span>
-          </div>
-          <div style="padding-left: 10px;">
-            <label class="radio-inline"><input type="radio" name="secure-radio-${option.id}" id="confirm-${option.id}" ><i class='fa fa-question' title="Demander confirmation"></i></label>
-            <label class="radio-inline"><input type="radio" name="secure-radio-${option.id}" id="secure-${option.id}"  ><i class='fa fa-fingerprint' title="Sécuriser avec empreinte digitale"></i></label>
-            <label class="radio-inline"><input type="radio" name="secure-radio-${option.id}" id="pwd-${option.id}"     ><i class='mdi mdi-numeric' title="Sécuriser avec un code"></i></label>
-            <label class="radio-inline"><input type="radio" name="secure-radio-${option.id}" id="none-${option.id}"  checked   >Aucun</label>
-          </div>
-        </div>
-      </div>
-
-
-    </div>
-    </div></li>`;
-    } else if (option.category == "choicesList") {
-      curOption += `<div class='input-group'>`;
-
-      option.choices.forEach(v => {
-        curOption += `<label class="checkbox-inline">
-      <input type="checkbox" class="eqLogicAttr" id="${v.id}-jc-checkbox" />${v.text}
-      </label>`;
-      });
-
-      curOption += `</div></div></div></li>`;
-    } else {
-      return;
-    }
-
-    items.push(curOption);
-
-  });
-
-  //More infos
-  if (!["widgets-summary", "room", "favorites"].includes(widget.type)) {
-    moreDiv = `<li><div class='form-group'>
-      <label class='col-xs-3 '>Ajouter des infos</label>
-      <div class='col-xs-9'>
-      <div class="description">Permet d'ajouter des infos utilisables dans les images sous conditions, nom et/ou sous-titre</div>
-      <div class='input-group'>
-      <span class="input-group-btn">
-        <a class="btn btn-default roundedRight" onclick="addMoreCmd()"><i class="fas fa-plus-square">
-        </i> Ajouter une commande</a>
-          </span>
-          </div>
-        <div id="moreInfos-div"></div>
-      </div></div></li>`;
-    items.push(moreDiv);
-  }
-
-  //Details access
-  option = `<li><div class='form-group'>
-    <label class='col-xs-3 '>Bloquer vue détails</label>
-    <div class='col-xs-9'><div class='input-group'><input type="checkbox" style="width:150px;" id="blockDetail-input" ></div></div></div></li>`;
-  items.push(option);
-
-  $("#widgetOptions").html(items.join(""));
-  loadSortable('all');
-}
-
-
-function imagePicker(elm) {
-  var newElt = $(elm).nextAll("a[data-id^='icon-']:first");
-
-  getIconModal({ title: "Choisir une icône ou une image", withIcon: "1", withImg: "1", icon: htmlToIcon(newElt.children().first()), elt: newElt }, (result, _params) => {
-    $(_params.elt).html(iconToHtml(result));
-  });
-}
-
-
-function removeImage(elm) {
-  $(elm).empty();
-}
-
-
-
-function removeCmd(id) {
-  $("#" + id + "-input").attr('value', '');
-  $("#" + id + "-input").val('');
-  $("#" + id + "-input").attr('cmdId', '');
-}
 
 function selectCmd(name, type, subtype, value) {
   var cmd = { type: type }
@@ -1384,13 +215,7 @@ function selectCmd(name, type, subtype, value) {
   })
 }
 
-$("#widgetOptions").on('change', '.needRefresh', function () {
-  var info = $(this).val();
-  var name = $(this).attr('id').replace('-input', '');
-  var configlink = $(this).attr('configlink') || 'undefined';
-  refreshCmdData(name, info, configlink);
 
-});
 
 
 
@@ -1473,37 +298,10 @@ function refreshCmdData(name, id, value) {
 }
 
 
-
-
-
-function selectScenario(name) {
-  jeedom.scenario.getSelectModal({}, function (result) {
-    $("#" + name + "-input").attr('value', result.human);
-    $("#" + name + "-input").val(result.human);
-    $("#" + name + "-input").attr('scId', result.id);
-    if ($("#name-input").val() == "") {
-      getScenarioHumanName({
-        id: name,
-        error: function (error) { },
-        success: function (data) {
-          data.forEach(sc => {
-            if (sc['id'] == result.id) {
-              $("#name-input").val(sc.name);
-            }
-          })
-        }
-      });
-      $("#name-input").val(result.name);
-
-      previousData = $('#tags-scenario-input').val() || '';
-      $('#tags-scenario-input').val(previousData);
-      $('#optionScenario').css('display', 'block');
-    }
-  })
-}
-
-function subtitleSelected() {
-  if ($("#subtitle-input").val() == 'custom') {
+// TODO
+// call twice !?
+$("body").on('change', '.subtitleSelected', function (e) {
+  if ($(this).val() == 'custom') {
     $("#subtitle-input-value").show();
     $("#subtitle-select").show();
   } else {
@@ -1511,7 +309,7 @@ function subtitleSelected() {
     $("#subtitle-select").hide();
     $("#subtitle-input-value").val($("#subtitle-input").val());
   }
-}
+});
 
 function refreshWidgetOption() {
   curOption = "";
@@ -1602,9 +400,9 @@ function refreshCmdListOption(optionsJson) {
       curOption += `<div class="col-lg-12 form-group">`;
       curOption += `
               <div>
-                      <a class="btn btn-success roundedRight" onclick="imagePicker(this)"><i class="fas fa-check-square">
+                      <a class="btn btn-success roundedRight imagePicker"><i class="fas fa-check-square">
                       </i> Icône </a>
-                      <a class="jcCmdListOptions" data-id="icon-${item.id}" data-index="${item.index}" onclick="removeImage(this)">${iconToHtml(item.image)}</a>
+                      <a class="jcCmdListOptions removeImage" data-id="icon-${item.id}" data-index="${item.index}">${iconToHtml(item.image)}</a>
               </div>`;
       curOption += `</div>`;
     }
@@ -1657,200 +455,6 @@ function refreshCmdListOption(optionsJson) {
 
   })
 }
-
-
-$("#widgetOptions").on('focusin', '.jcCmdListOptionsCommand', function () {
-  $(this).attr('data-value-focusin', $(this).val());
-})
-
-$("#widgetOptions").on('focusout', '.jcCmdListOptionsCommand', function () {
-  var previousData = $(this).attr('data-value-focusin');
-  if (previousData != $(this).val()) {
-
-    var elt = $(this);
-    var currentId = elt.attr('data-cmd-id');
-    var currentIndex = elt.attr('data-index');
-
-    getCmd({
-      id: elt.val(),
-      error: function (error) {
-        $('#widget-alert').showAlert({ message: error.result, level: 'danger' });
-      },
-      success: function (data) {
-        $('#widget-alert').hideAlert();
-
-        // remove options of the previous cmd
-        elt.parent().siblings().each(function () {
-          if ($(this).find('.jcCmdListOptions').length !== 0) {
-            $(this).remove();
-          }
-        });
-
-        // create options for the new cmd selected
-        newItem = { id: data.result.id, subtype: data.result.subType, index: currentIndex, name: data.result.name };
-        divListOption = getCmdOptions(newItem);
-        elt.parent().parent().append(divListOption);
-
-        // update cmdCat
-        cmdCat.forEach(item => {
-          if (item.id == currentId && item.index == currentIndex) {
-            item.id = data.result.id;
-            item.name = data.result.humanName;
-            item.subtype = data.result.subType;
-          }
-        })
-        saveCmdList();
-
-      }
-    });
-
-  }
-
-  $(this).attr('data-value-focusin', '');
-})
-
-
-function getCmdOptions(item) {
-
-  curOption = '';
-
-  var d = new Date().getTime();
-  var rand = Math.floor(Math.random() * 10000) * Math.floor(Math.random() * 10000);
-  var timestamp = d + '' + rand;
-  var customUid = 'cmd' + item.id + '____' + timestamp + '____';
-
-  // create dynamic var
-  if ('options' in item) {
-    for (const [key, value] of Object.entries(item.options)) {
-      eval('var option' + key.charAt(0).toUpperCase() + key.slice(1) + '= "' + value + '";');
-    }
-  }
-
-
-  if (item.subtype == 'select') {
-    var optionSelect = optionSelect || '';
-
-    getCmdDetail({ id: item.id, item: item, optionSelect: optionSelect }, function (_result, _param) {
-
-      var selectOptions = [];
-      $.each(_result.configuration.listValue.split(';'), function (key, val) {
-        var myData = val.split('|');
-        if (myData.length == 1) {
-          var value = text = myData[0];
-        }
-        else {
-          var value = myData[0];
-          var text = myData[1];
-        }
-        selectOptions.push(`<option value="${value}">${text}</option>`);
-      });
-
-      $('.jcCmdListOptions[data-uid=' + customUid + '][data-l1key=options][data-l2key=select]').html(selectOptions.join(""));
-
-      var optionSelect = _param.optionSelect || '';
-
-      $('.jcCmdListOptions[data-uid=' + customUid + '][data-l1key=options][data-l2key=select]').value(optionSelect);
-
-
-    });
-
-    curOption = `
-        <div class="input-group input-group-sm" style="width: 100%">
-            <span class="input-group-addon roundedLeft" style="width: 100px">Choix</span>
-            <select class="form-control input-sm jcCmdListOptions roundedRight" data-l1key="options" data-l2key="select" data-uid="${customUid}" data-id="select-${item.id}" data-index="${item.index}">
-            </select>
-        </div>
-       `;
-
-  }
-
-
-  if (item.subtype == 'message') {
-
-    var optionTitle = optionTitle || '';
-    var optionMessage = optionMessage || '';
-
-    curOption = `
-        <div class="input-group input-group-sm" style="width: 100%">
-            <span class="input-group-addon roundedLeft" style="width: 100px">Titre</span>
-            <input style="width:240px;" class='input-sm form-control roundedRight title jcCmdListOptions' type="string" data-id="title-${item.id}" data-index="${item.index}" value="${optionTitle}" />
-        </div>
-        <div class="input-group input-group-sm" style="width: 100%">
-            <span class="input-group-addon roundedLeft" style="width: 100px">Message</span>
-            <textarea class="message form-control ta_autosize jcCmdListOptions" data-l1key="options" data-l2key="message" rows="1" style="resize:vertical;"  data-id="message-${item.id}" data-index="${item.index}" data-uid="${customUid}">${optionMessage}</textarea>
-            <span class="input-group-addon hasBtn roundedRight">
-              <button class="btn btn-default roundedRight listEquipementInfo" type="button" tooltip="Sélectionner la commande" data-cmd_id="${item.id}" data-index="${item.index}" data-uid="${customUid}" ><i class="fas fa-list-alt"></i></button>
-            </span>
-
-        <script>
-          $('.listEquipementInfo[data-uid=${customUid}]').on('click', function() {
-              jeedom.cmd.getSelectModal({cmd: {type: 'info'}}, function(result) {
-                $('.jcCmdListOptions[data-l1key=options][data-l2key=message][data-uid=${customUid}]').atCaret('insert', result.human);
-              });
-          });
-
-        </script>
-        </div>`;
-  }
-
-  if (item.subtype == 'slider') {
-    var optionSlider = optionSlider || '';
-
-    curOption = `
-        <div class="input-group input-group-sm" style="width: 100%">
-            <span class="input-group-addon roundedLeft" style="width: 100px">Valeur</span>
-            <textarea class="message form-control ta_autosize jcCmdListOptions" data-l1key="options" data-l2key="message" rows="1" style="resize:vertical;"  data-id="slider-${item.id}" data-index="${item.index}" data-uid="${customUid}">${optionSlider}</textarea>
-            <span class="input-group-addon hasBtn roundedRight">
-              <button class="btn btn-default roundedRight listEquipementInfo" type="button" tooltip="Sélectionner la commande" data-cmd_id="${item.id}" data-index="${item.index}" data-uid="${customUid}"><i class="fas fa-list-alt"></i></button>
-            </span>
-
-        <script>
-          $('.listEquipementInfo[data-uid=${customUid}]').on('click', function() {
-              jeedom.cmd.getSelectModal({cmd: {type: 'info'}}, function(result) {
-                $('.jcCmdListOptions[data-l1key=options][data-l2key=message][data-uid=${customUid}]').atCaret('insert', result.human);
-              });
-          });
-        </script>
-        </div>`;
-  }
-
-  if (item.subtype == 'color') {
-    var optionColor = optionColor || '';
-
-    curOption = `
-        <div class="input-group input-group-sm" style="width: 100%">
-            <span class="input-group-addon roundedLeft" style="width: 100px">Couleur</span>
-            <input type="color" class="form-control input-sm cursor colorChooser" data-uid="${customUid}" style="width: 20%; display: inline-block;" value="${optionColor}">
-            <input class="jcCmdListOptions form-control input-sm" data-l1key="options" data-l2key="color"  data-uid="${customUid}" style="width: 80%; display: inline-block;" placeholder="Couleur (hexa)" data-id="color-${item.id}" data-index="${item.index}" value="${optionColor}">
-            <span class="input-group-btn">
-              <button class="btn btn-default listEquipementInfo roundedRight" type="button" tooltip="Sélectionner la commande" data-uid="${customUid}" data-index="${item.index}" data-cmd_id="${item.id}"><i class="fas fa-list-alt"></i></button>
-            </span>
-
-        <script>
-          $('.listEquipementInfo[data-uid=${customUid}]').off('click').on('click', function () {
-            var el = $(this);
-            jeedom.cmd.getSelectModal({cmd: {type: 'info'}}, function (result) {
-              $('.jcCmdListOptions[data-uid=${customUid}][data-l1key=options][data-l2key=color]').value(result.human);
-            });
-          });
-          $('.colorChooser[data-uid=${customUid}]').off('change').on('change', function () {
-            $('.jcCmdListOptions[data-uid=${customUid}][data-l1key=options][data-l2key=color]').value($(this).value())
-          });
-        </script>
-        </div>`;
-  }
-
-  curOption += `<div class="input-group input-group-sm" style="width: 100%">
-                            <span class="input-group-addon roundedLeft" style="width: 100px">Nom</span>
-                            <input style="width:240px;" class='input-sm form-control roundedRight title jcCmdListOptions' data-id="custom-name-${item.id}" data-index="${item.index}" value='${item.name || ""}' >
-                        </div>`;
-
-
-  return curOption;
-
-
-}
-
 
 
 function saveCmdList() {
@@ -1914,41 +518,6 @@ function deleteCmdOption(elm, optionsJson) {
   refreshCmdListOption(optionsJson);
 }
 
-/*
-function upCmdOption(elm, optionsJson) {
-  saveCmdList();
-
-  var id = $(elm).data('id');
-  var currentIndex = $(elm).data('index');
-
-  var cmdToMove = cmdCat.find(i => i.id == parseInt(id) && i.index == currentIndex);
-  var index = parseInt(cmdToMove.index);
-  if (index == 0) {
-    return;
-  }
-  var otherCmd = cmdCat.find(i => i.index == index - 1);
-  cmdToMove.index = index - 1;
-  otherCmd.index = index;
-  refreshCmdListOption(optionsJson);
-}
-
-function downCmdOption(elm, optionsJson) {
-  saveCmdList();
-
-  var id = $(elm).data('id');
-  var currentIndex = $(elm).data('index');
-
-  var cmdToMove = cmdCat.find(i => i.id == parseInt(id) && i.index == currentIndex );
-  var index = parseInt(cmdToMove.index);
-  if (index == getMaxIndex(cmdCat)) {
-    return;
-  }
-  var otherCmd = cmdCat.find(i => i.index == index + 1);
-  cmdToMove.index = index + 1;
-  otherCmd.index = index;
-  refreshCmdListOption(optionsJson);
-}
-*/
 
 //More Infos
 
@@ -2026,11 +595,6 @@ function refreshInfoSelect() {
   refreshStrings();
 }
 
-function colorDefined(c, id) {
-  // console.log('id', id)
-  $("#" + id + "-input").val(c.value);
-}
-
 function infoSelected(value, el) {
   let inputId = $(el).parent().attr("input")
   //$("#"+inputId).val( $("#"+inputId).val() + value);
@@ -2047,19 +611,7 @@ function refreshStrings() {
   $("#subtitle-input-value").val(idToHuman($("#subtitle-input-value").val(), infoCmd));
 }
 
-function idToHuman(string, infos) {
-  let result = string;
-  if (typeof (string) != "string") { return string; }
-  const match = string.match(/#.*?#/g);
-  if (!match) { return string; }
-  match.forEach(item => {
-    const info = infos.find(i => i.id == item.replace(/\#/g, ""));
-    if (info && info.human != '' && info.human != undefined) {
-      result = result.replace(item, info.human);
-    }
-  });
-  return result;
-}
+
 
 // Image condition list
 function refreshImgListOption(dataType = 'widget') {
@@ -2094,22 +646,6 @@ function refreshImgListOption(dataType = 'widget') {
   });
 
   imgCat.forEach(item => {
-    /*
-    curOption += `
-    <div data-id="${item.index}" class='input-group jcImgListMovable'>
-      Si
-      <input style="width:385px;height:31px;margin-left:5px" class=' roundedLeft' index="${item.index}" id="cond-input-${item.index}" value="${item.condition}"
-       onchange="setCondValue(this, 'imgList')" />
-       <a class='btn btn-default btn-sm cursor bt_selectTrigger' style=";margin-right:10px;" tooltip='Ajouter une commande' onclick="selectInfoCmd('#cond-input-${item.index}', 'imgList');">
-                    <i class='fas fa-list-alt'></i></a>
-      <a class="btn btn-success roundedRight" index="${item.index}" onclick="imagePicker(this)"><i class="fas fa-plus-square">
-      </i> Image </a>
-      <a data-id="icon-div" id="icon-div-${item.index}" onclick="removeImage(this)">${iconToHtml(item.image)}</a>
-      <i class="mdi mdi-arrow-up-down-bold" title="Déplacer" style="color:rgb(80, 120, 170);font-size:24px;margin-right:10px;margin-left:10px;cursor:grab!important;" aria-hidden="true"></i>
-      <i class="mdi mdi-minus-circle" style="color:rgb(185, 58, 62);font-size:24px;" aria-hidden="true" onclick="deleteImgOption('${item.index}');"></i>
-      </div>
-    `;
-    */
     curOption += `
 		<div data-id="${item.index}" class='input-group jcImgListMovable'>
 			Si
@@ -2126,9 +662,9 @@ function refreshImgListOption(dataType = 'widget') {
       });
     }
     curOption += `</ul></div>`;
-    curOption += `<a class="btn btn-success roundedRight" index="${item.index}" onclick="imagePicker(this)"><i class="fas fa-plus-square">
+    curOption += `<a class="btn btn-success roundedRight imagePicker" index="${item.index}"><i class="fas fa-plus-square">
     </i> Image </a>
-    <a data-id="icon-div" id="icon-div-${item.index}" onclick="removeImage(this)">${iconToHtml(item.image)}</a>
+    <a data-id="icon-div" id="icon-div-${item.index}" class="removeImage">${iconToHtml(item.image)}</a>
     <i class="mdi mdi-arrow-up-down-bold" title="Déplacer" style="color:rgb(80, 120, 170);font-size:24px;margin-right:10px;margin-left:10px;cursor:grab!important;" aria-hidden="true"></i>
     <i class="mdi mdi-minus-circle" style="color:rgb(185, 58, 62);font-size:24px;" aria-hidden="true" onclick="deleteImgOption('${item.index}');"></i>
     </div>` ;
@@ -2246,33 +782,6 @@ function deleteImgOption(id) {
   refreshImgListOption();
 }
 
-/*
-function upImgOption(id) {
-  saveImgOption();
-  var imgToMove = imgCat.find(i => i.index == parseInt(id));
-  var index = parseInt(imgToMove.index);
-  if (index == 0) {
-    return;
-  }
-  var otherImg = imgCat.find(i => i.index == index - 1);
-  imgToMove.index = index - 1;
-  otherImg.index = index;
-  refreshImgListOption();
-}
-
-function downImgOption(id) {
-  saveImgOption();
-  var imgToMove = imgCat.find(i => i.index == parseInt(id));
-  var index = parseInt(imgToMove.index);
-  if (index == getMaxIndex(imgCat)) {
-    return;
-  }
-  var otherImg = imgCat.find(i => i.index == index + 1);
-  imgToMove.index = index + 1;
-  otherImg.index = index;
-  refreshImgListOption();
-}
-*/
 
 function addWidgetOption(choices) {
   var widgets = choices.split(".");
@@ -2295,153 +804,8 @@ function deleteWidgetOption(id) {
   refreshWidgetOption();
 }
 
-/*
-function upWidgetOption(id) {
-  var widgetToMove = widgetsCat.find(i => i.id == parseInt(id));
-  var index = parseInt(widgetToMove.index);
-  if (index == 0) {
-    return;
-  }
-  var otherWidget = widgetsCat.find(i => i.index == index - 1);
-  widgetToMove.index = index - 1;
-  otherWidget.index = index;
-  refreshWidgetOption();
-}
 
-function downWidgetOption(id) {
-  var widgetToMove = widgetsCat.find(i => i.id == parseInt(id));
-  var index = parseInt(widgetToMove.index);
-  if (index == getMaxIndex(widgetsCat)) {
-    return;
-  }
-  var otherWidget = widgetsCat.find(i => i.index == index + 1);
-  widgetToMove.index = index + 1;
-  otherWidget.index = index;
-  refreshWidgetOption();
-}
-*/
-
-function getHumanName(cmdId) {
-  myCmd = allJeedomData.find(i => i.id == cmdId);
-  myHumanName = myCmd?.humanName ? '#' + myCmd.humanName + '#' : '';
-  return myHumanName;
-}
-
-function getHumanName_old(_params) {
-  var params = $.extend({}, jeedom.private.default_params, {}, _params || {});
-
-  var paramsAJAX = jeedom.private.getParamsAJAX(params);
-  paramsAJAX.url = 'core/ajax/cmd.ajax.php';
-  paramsAJAX.data = {
-    action: 'getHumanCmdName',
-    id: _params.id
-  };
-  $.ajax(paramsAJAX);
-}
-
-
-function getCmdIdFromHumanName(_params, _callback) {
-  if (typeof _params.alert == 'undefined') {
-    _params.alert = '#div_alert';
-  }
-
-  $.post({
-    url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-    data: {
-      action: 'humanReadableToCmd',
-      human: _params.stringData
-    },
-    cache: false,
-    dataType: 'json',
-    async: false,
-    success: function (data) {
-      if (data.state != 'ok') {
-        $(_params.alert).showAlert({
-          message: data.result,
-          level: 'danger'
-        });
-      }
-      else {
-        if ('function' == typeof (_callback)) {
-          _callback(data.result, _params);
-        }
-      }
-    }
-  });
-}
-
-function getHumanNameFromCmdId(_params, _callback) {
-  if (typeof _params.alert == 'undefined') {
-    _params.alert = '#div_alert';
-  }
-
-  $.post({
-    url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-    data: {
-      action: 'cmdToHumanReadable',
-      strWithCmdId: _params.cmdIdData
-    },
-    cache: false,
-    dataType: 'json',
-    async: false,
-    success: function (data) {
-      if (data.state != 'ok') {
-        $(_params.alert).showAlert({
-          message: data.result,
-          level: 'danger'
-        });
-      }
-      else {
-        if ('function' == typeof (_callback)) {
-          _callback(data.result, _params);
-        }
-      }
-    }
-  });
-}
-
-function getCmd({ id, error, success }) {
-  $.post({
-    url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-    data: { 'action': 'getCmd', 'id': id },
-    cache: false,
-    success: function (cmdData) {
-      jsonData = JSON.parse(cmdData);
-      if (jsonData.state == 'ok') {
-        success && success(jsonData);
-      } else {
-        error && error(jsonData);
-      }
-    }
-  });
-}
-
-function jeedomIconToIcon(html) {
-  if (html.startsWith("<i ")) {
-    let tag1 = html.split("\"")[1].split(" ")[0];
-    let tag2 = html.split("\"")[1].split(" ")[1];
-    if (tag1 == 'icon') {
-      return { source: 'jeedom', name: tag2 };
-    } else if (tag1.startsWith('fa')) {
-      return { source: 'fa', prefix: tag1, name: tag2.replace("fa-", "") };
-    }
-  }
-}
-
-function getScenarioHumanName(_params) {
-  var params = $.extend({}, jeedom.private.default_params, {}, _params || {});
-
-  var paramsAJAX = jeedom.private.getParamsAJAX(params);
-  paramsAJAX.url = 'core/ajax/scenario.ajax.php';
-  paramsAJAX.data = {
-    action: 'all',
-    id: _params.id
-  };
-  $.ajax(paramsAJAX);
-}
-
-
-function saveWidget() {
+$(".widgetMenu .saveWidget").click(function () {
   $('#widget-alert').hideAlert();
 
   try {
@@ -2720,53 +1084,29 @@ function saveWidget() {
     console.error(error);
   }
 
-}
+});
 
-function getCustomParamUrl(url, vars) {
 
-  for (var i in vars) {
-    if (i != 'jcOrderBy' && i != 'jcFilter' && i != 'jcSearch') {
-      url += i + '=' + vars[i].replace('#', '') + '&'
-    }
-  }
 
-  var widgetFilter = $("#widgetTypeSelect option:selected").val();
-  if (widgetFilter != 'none') {
-    url += '&jcFilter=' + widgetFilter
-  }
-
-  var widgetOrder = $("#widgetOrder option:selected").val();
-  if (widgetOrder != 'none') {
-    url += '&jcOrderBy=' + widgetOrder
-  }
-
-  var widgetSearch = $("#in_searchWidget").val().trim();
-  if (widgetSearch != '') {
-    url += '&jcSearch=' + widgetSearch
-  }
-
-  return url;
-
-}
-
-function hideWidget() {
+$(".widgetMenu .hideWidget").click(function () {
   $("#widgetModal").dialog('destroy').remove();
-}
+});
 
-function duplicateWidget() {
+$(".widgetMenu .duplicateWidget").click(function () {
+
   $('#widget-alert').hideAlert();
 
   $('#widgetOptions').attr('widget-id', '');
 
-  $('.widgetMenu .duplicateWidget').hide()
+  $(this).hide()
   $('.widgetMenu .removeWidget').hide()
   showAutoFillWidgetCmds();
   $('#widget-alert').showAlert({ message: 'Vous êtes sur le widget dupliqué, réalisez (ou non) vos modifications. Dans tous les cas, pensez à sauvegarder !', level: 'success' });
   // $('.widgetMenu .saveWidget').attr('exit-attr', 'true');
 
-}
+});
 
-function removeWidget(itemId) {
+$(".widgetMenu .removeWidget").click(function () {
   var warning = "<i source='md' name='alert-outline' style='color:#ff0000' class='mdi mdi-alert-outline'></i>";
 
   if (itemId == undefined) {
@@ -2838,7 +1178,7 @@ function removeWidget(itemId) {
     })
   });
 
-}
+});
 
 
 function getWidgetPath(id) {
@@ -2884,142 +1224,6 @@ function getRoomName(id) {
   }
 }
 
-function getMaxIndex(array) {
-  var maxIndex = -1;
-  array.forEach(item => {
-    if (item.index > maxIndex) {
-      maxIndex = item.index;
-    }
-  });
-  return maxIndex;
-}
-
-
-function getMaxId(array, defaut = -1) {
-  var maxId = defaut;
-  array.forEach(item => {
-    if (item.id > maxId) {
-      maxId = item.id;
-    }
-  });
-  return maxId;
-}
-
-function parseString(string, infos) {
-  let result = string;
-  if (typeof (string) != "string") { return string; }
-  const match = string.match(/#.*?#/g);
-  if (!match) { return string; }
-  match.forEach(item => {
-    const info = infos.find(i => i.human == item);
-    if (info) {
-      result = result.replace(item, "#" + info.id + "#");
-    }
-  });
-  return result;
-}
-
-
-function updateOrderWidget() {
-
-  var type = $("#widgetOrder").val();
-
-  var vars = getUrlVars()
-  var url = 'index.php?'
-
-  url = getCustomParamUrl(url, vars);
-
-  loadPage(url)
-
-}
-
-
-$('#widgetOrder_NOTWORKING').on('change', function () {
-  var type = $("#widgetOrder").val();
-
-  $.post({
-    url: "plugins/JeedomConnect/core/ajax/jeedomConnect.ajax.php",
-    data: {
-      action: 'orderWidget',
-      orderBy: type
-    },
-    cache: false,
-    dataType: 'json',
-    async: false,
-    success: function (data) {
-      if (data.state != 'ok') {
-        $('#div_alert').showAlert({
-          message: data.result,
-          level: 'danger'
-        });
-      }
-      else {
-        // $('.eqLogicThumbnailContainer').hide() ;
-        $('#widgetsList-div').html(data.result.widgets);
-        // $('.eqLogicThumbnailContainer').show() ;
-        $('.eqLogicThumbnailContainer').packery();
-
-      }
-    }
-  });
-
-
-
-});
-
-
-$('#widgetTypeSelect').on('change', function () {
-  var typeSelected = this.value;
-
-  $('.widgetDisplayCard').show();
-  if (typeSelected != 'none') {
-    $('.widgetDisplayCard').not("[data-widget_type=" + typeSelected + "]").hide();
-  }
-
-  var widgetSearch = $("#in_searchWidget").val().trim();
-  if (widgetSearch != '') {
-    $('#in_searchWidget').keyup()
-  }
-
-  $('.eqLogicThumbnailContainer').packery();
-
-  updateWidgetCount();
-
-});
-
-
-$('body').off('click', '.toggle-password').on('click', '.toggle-password', function () {
-  $(this).toggleClass("fa-eye fa-eye-slash");
-  var input = $("#actionPwd");
-  if (input.attr("type") === "password") {
-    input.attr("type", "text");
-  } else {
-    input.attr("type", "password");
-  }
-
-});
-
-var originalPwd = null;
-$("#actionPwd").focusin(function () {
-  if (originalPwd === null) {
-    originalPwd = $(this).val();
-  }
-});
-
-
-function saveEqLogic(_eqLogic) {
-  if (!isset(_eqLogic.configuration)) {
-    _eqLogic.configuration = {};
-  }
-
-  currentPwd = $("#actionPwd").val();
-  if (originalPwd !== null && originalPwd != currentPwd) {
-    _eqLogic.configuration.pwdChanged = 'true';
-  }
-
-  return _eqLogic;
-
-}
 
 
 function getCmdDetail(_params, _callback) {
@@ -3054,34 +1258,6 @@ function getCmdDetail(_params, _callback) {
   };
   $.ajax(paramsAJAX);
 };
-
-$('#widgetsList-div').on('change', function () {
-  updateWidgetCount()
-})
-
-function updateWidgetCount() {
-
-  var nbVisible = $('.widgetDisplayCard:visible').length;
-  var nbTotal = $('.widgetDisplayCard').length;
-
-  var text = (nbTotal != nbVisible) ? nbVisible + "/" + nbTotal : nbTotal;
-
-  $('#coundWidget').text("(" + text + ")");
-}
-
-
-$('#eraseFilterChoice').off('click').on('click', function () {
-  var vars = getUrlVars()
-  var url = 'index.php?'
-  for (var i in vars) {
-    if (i != 'id' && i != 'saveSuccessFull' && i != 'removeSuccessFull' &&
-      i != 'jcOrderBy' && i != 'jcSearch' && i != 'jcFilter') {
-      url += i + '=' + vars[i].replace('#', '') + '&'
-    }
-  }
-
-  loadPage(url)
-})
 
 
 function selectInfoCmd(elt, conf) {
@@ -3173,6 +1349,336 @@ $('body').off('click', '.removeParent').on('click', '.removeParent', function ()
   $(this).parent().remove();
 });
 
+
+/**
+ * *************
+ * TOP ACTION BUTTON
+ * *************
+ */
+
+
+$('.eqLogicAction[data-action=addWidget]').off('click').on('click', function () {
+  getWidgetModal({ title: "Configuration du widget", removeAction: false, exit: true });
+})
+
+$('.eqLogicAction[data-action=showError]').off('click').on('click', function () {
+
+  var hide = ($('#spanWidgetErreur').text() == 'Erreur') ? true : false;
+  if (hide) {
+    $('.widgetDisplayCard').not(".hasError").hide();
+  }
+  else {
+    $('.widgetDisplayCard').show();
+  }
+
+
+  var typeSelected = $('#widgetTypeSelect').val();
+
+  if (typeSelected != 'none') {
+    $('.widgetDisplayCard').not("[data-widget_type=" + typeSelected + "]").hide();
+  }
+
+
+  $('.eqLogicThumbnailContainer').packery();
+
+  if (hide) {
+    $('#spanWidgetErreur').text('Tous');
+    $('.eqLogicAction[data-action=showError]').css('color', 'grey');
+  }
+  else {
+    $('#spanWidgetErreur').text('Erreur');
+    $('.eqLogicAction[data-action=showError]').css('color', 'red');
+  }
+  updateWidgetCount()
+
+})
+
+$('.eqLogicAction[data-action=addWidgetBulk]').off('click').on('click', function () {
+  $("#widgetModal").dialog('destroy').remove();
+  $('body').append('<div id="widgetModal"></div>');
+  $('#widgetModal').dialog({
+    title: "{{Ajout de widgets en masse}}",
+    width: 0.95 * $(window).width(),
+    height: 0.8 * $(window).height(),
+    modal: true,
+    closeText: ''
+  });
+  $('#widgetModal').load('index.php?v=d&plugin=JeedomConnect&modal=assistant.widgetBulkModal.JeedomConnect').dialog('open');
+})
+
+
+$('.eqLogicAction[data-action=showCommunity]').off('click').on('click', function () {
+  // $('.pluginInfo').toggle("slide", { direction: "right" }, 1000);
+  getSimpleModal({
+    title: "Forum",
+    width: 0.5 * $(window).width(),
+    fields: [{
+      type: "string",
+      value: $('.txtInfoPlugin').html()
+    },
+    {
+      type: "string",
+      id: "infoPluginModal",
+      value: $('.infoPlugin').html()
+    }],
+    buttons: {
+      "Fermer": function () {
+        $('#simpleModalAlert').hide();
+        $(this).dialog("close");
+      },
+      "Copier": function () {
+        copyDivToClipboard('#infoPluginModal', true)
+      }
+    }
+  }, function (result) { });
+
+
+});
+
+$('.eqLogicAction[data-action=showSummary]').off('click').on('click', function () {
+  $('body').append('<div id="widgetSummaryModal"></div>');
+  $('#widgetSummaryModal').dialog({
+    title: "{{Synthèse globale des widgets}}",
+    autoOpen: false,
+    modal: true,
+    closeText: '',
+    width: 0.9 * $(window).width(),
+    height: 0.8 * $(window).height(),
+    closeOnEscape: false,
+    // open: function(event, ui) { $(".ui-dialog-titlebar-close").hide(); },
+    close: function (ev, ui) { check_before_closing(); }
+  });
+  $('#widgetSummaryModal').load('index.php?v=d&plugin=JeedomConnect&modal=assistant.widgetSummary.JeedomConnect').dialog('open');
+})
+
+$('.eqLogicAction[data-action=showEquipmentSummary]').off('click').on('click', function () {
+  $('body').append('<div id="equipmentSummaryModal"></div>');
+  $('#equipmentSummaryModal').dialog({
+    title: "{{Synthèse des équipements JC}}",
+    autoOpen: false,
+    modal: true,
+    closeText: '',
+    width: 0.9 * $(window).width(),
+    height: 0.8 * $(window).height(),
+    closeOnEscape: false,
+    // open: function(event, ui) { $(".ui-dialog-titlebar-close").hide(); },
+    close: function (ev, ui) { check_before_closing(); }
+  });
+  $('#equipmentSummaryModal').load('index.php?v=d&plugin=JeedomConnect&modal=assistant.equipmentSummary.JeedomConnect').dialog('open');
+})
+
+$('.eqLogicAction[data-action=showNotifAll]').off('click').on('click', function () {
+  $('body').append('<div id="notifAllModal"></div>');
+  $('#notifAllModal').dialog({
+    title: "{{Configuration de la notification de \"tous\" les équipements}}",
+    autoOpen: false,
+    modal: true,
+    closeText: '',
+    width: 0.7 * $(window).width(),
+    height: 0.8 * $(window).height(),
+    closeOnEscape: false
+  });
+  $('#notifAllModal').load('index.php?v=d&plugin=JeedomConnect&modal=assistant.notifAll.JeedomConnect').dialog('open');
+})
+
+// Start Generic Types
+// HACK Remove when gentype config in plugin is not needed anymore
+function gotoGenTypeConfig() {
+  $('#md_modal').dialog({ title: "{{Objets / Pièces}}" });
+  $('#md_modal').load('index.php?v=d&plugin=JeedomConnect&modal=gentype.objects').dialog('open');
+}
+
+$('.eqLogicAction[data-action=gotoGenTypeConfig]').off('click').on('click', function () {
+  gotoGenTypeConfig();
+})
+// End Generic Types
+
+// ------------- END TOP ACTION BUTTON
+
+
+
+/**
+ * *************
+ * SEARCH, ORDER & FILTER BAR
+ * *************
+ */
+
+$('#widgetsList-div').on('change', function () {
+  updateWidgetCount()
+})
+
+function updateWidgetCount() {
+
+  var nbVisible = $('.widgetDisplayCard:visible').length;
+  var nbTotal = $('.widgetDisplayCard').length;
+
+  var text = (nbTotal != nbVisible) ? nbVisible + "/" + nbTotal : nbTotal;
+
+  $('#coundWidget').text("(" + text + ")");
+}
+
+$('#widgetTypeSelect').on('change', function () {
+  var typeSelected = this.value;
+
+  $('.widgetDisplayCard').show();
+  if (typeSelected != 'none') {
+    $('.widgetDisplayCard').not("[data-widget_type=" + typeSelected + "]").hide();
+  }
+
+  var widgetSearch = $("#in_searchWidget").val().trim();
+  if (widgetSearch != '') {
+    $('#in_searchWidget').keyup()
+  }
+
+  $('.eqLogicThumbnailContainer').packery();
+
+  updateWidgetCount();
+
+});
+
+$('#eraseFilterChoice').off('click').on('click', function () {
+  var vars = getUrlVars()
+  var url = 'index.php?'
+  for (var i in vars) {
+    if (i != 'id' && i != 'saveSuccessFull' && i != 'removeSuccessFull' &&
+      i != 'jcOrderBy' && i != 'jcSearch' && i != 'jcFilter') {
+      url += i + '=' + vars[i].replace('#', '') + '&'
+    }
+  }
+
+  loadPage(url)
+})
+
+if ($('#in_searchWidget').val() != '') {
+  $('#in_searchWidget').keyup();
+}
+
+$('#bt_resetSearchWidget').on('click', function () {
+  $('#in_searchWidget').val('').keyup()
+})
+
+$('.updateOrderWidget').on('change', function () {
+
+  var type = $(this).val();
+
+  var vars = getUrlVars()
+  var url = 'index.php?'
+
+  url = getCustomParamUrl(url, vars);
+
+  loadPage(url)
+
+});
+
+function getCustomParamUrl(url, vars) {
+
+  for (var i in vars) {
+    if (i != 'jcOrderBy' && i != 'jcFilter' && i != 'jcSearch') {
+      url += i + '=' + vars[i].replace('#', '') + '&'
+    }
+  }
+
+  var widgetFilter = $("#widgetTypeSelect option:selected").val();
+  if (widgetFilter != 'none') {
+    url += '&jcFilter=' + widgetFilter
+  }
+
+  var widgetOrder = $("#widgetOrder option:selected").val();
+  if (widgetOrder != 'none') {
+    url += '&jcOrderBy=' + widgetOrder
+  }
+
+  var widgetSearch = $("#in_searchWidget").val().trim();
+  if (widgetSearch != '') {
+    url += '&jcSearch=' + widgetSearch
+  }
+
+  return url;
+}
+
+function sortWidgets() {
+  if (jcOrderBy === undefined) jcOrderBy = 'object';
+
+  allWidgetsDetail.sort(SortByName);
+  if (jcOrderBy == 'object') {
+    allWidgetsDetail.sort(SortByRoom);
+  } else if (jcOrderBy == 'type') {
+    allWidgetsDetail.sort(SortByType);
+  }
+}
+
+function SortByName(a, b) {
+  var aName = a.name?.toLowerCase();
+  var bName = b.name?.toLowerCase();
+  return ((aName < bName) ? -1 : ((aName > bName) ? 1 : 0));
+}
+
+function SortByType(a, b) {
+  var aType = a.type.toLowerCase();
+  var bType = b.type.toLowerCase();
+  return ((aType < bType) ? -1 : ((aType > bType) ? 1 : 0));
+}
+
+function SortByRoom(a, b) {
+
+  var aObj = ('room' in a) ? roomList.find(r => r.id == a.room) || undefined : undefined;
+  var bObj = ('room' in b) ? roomList.find(r => r.id == b.room) || undefined : undefined;
+
+  var aRoom = (aObj != undefined) ? aObj.name.toLowerCase() : 'AAAucun';
+  var bRoom = (bObj != undefined) ? bObj.name.toLowerCase() : 'AAAucun';
+
+  return ((aRoom < bRoom) ? -1 : ((aRoom > bRoom) ? 1 : 0));
+}
+
+$('#in_searchWidget').off('keyup').keyup(function () {
+  var search = $(this).value()
+  var widgetFilter = $("#widgetTypeSelect option:selected").val();
+
+  if (search == '') {
+    if (widgetFilter == 'none') {
+      $('.widgetDisplayCard').show()
+    }
+    else {
+      $('.widgetDisplayCard').each(function () {
+        widgetType = $(this).attr('data-widget_type');
+        if (widgetFilter == widgetType) {
+          $(this).closest('.widgetDisplayCard').show()
+        }
+      })
+    }
+    $('.eqLogicThumbnailContainer').packery()
+    updateWidgetCount();
+    return;
+  }
+
+
+  $('.widgetDisplayCard').hide()
+  search = normTextLower(search)
+  var text
+  var widgetId
+
+  $('.widgetDisplayCard').each(function () {
+    text = normTextLower($(this).children('.name').text())
+    widgetId = normTextLower($(this).attr('data-widget_id'))
+    widgetType = $(this).attr('data-widget_type');
+    if (text.indexOf(search) >= 0 || widgetId.indexOf(search) >= 0) {
+      if (widgetFilter == 'none' || widgetFilter == widgetType) {
+        $(this).closest('.widgetDisplayCard').show()
+      }
+    }
+  })
+  $('.eqLogicThumbnailContainer').packery()
+  updateWidgetCount();
+})
+
+// ------------- END SEARCH & FILTER BAR
+
+/**
+ * *************
+ * COMMUNITY INFO
+ * *************
+ */
+
 function copyDivToClipboard(myInput, addBacktick = false) {
   var initialText = $(myInput).html();
   if (addBacktick) {
@@ -3192,46 +1698,14 @@ function copyDivToClipboard(myInput, addBacktick = false) {
     $(myInput).html(initialText);
   }
 }
-
-updateWidgetCount();
-
-
-if ($('#in_searchWidget').val() != '') {
-  $('#in_searchWidget').keyup();
-}
+// ------------- END COMMUNITY INFO
 
 
-$('.eqLogicAttr[data-l1key=configuration][data-l2key=polling]').on('change', function () {
-
-  disableCheckbox()
-
-});
-
-$('.eqLogicAttr[data-l1key=configuration][data-l2key=useWs]').on('change', function () {
-
-  disableCheckbox()
-
-});
-
-function disableCheckbox() {
-  var useWs = $('.eqLogicAttr[data-l1key=configuration][data-l2key=useWs]');
-  var polling = $('.eqLogicAttr[data-l1key=configuration][data-l2key=polling]');
-
-  polling.attr('disabled', false);
-  useWs.attr('disabled', false);
-
-  if (useWs.prop("checked")) {
-    polling.prop("checked", false);
-    polling.attr('disabled', true);
-  }
-  else {
-    if (polling.prop("checked")) {
-      useWs.prop("checked", false);
-      useWs.attr('disabled', true);
-    }
-  }
-
-}
+/**
+ * *************
+ * WARNING COMMUNITY INFO
+ * *************
+ */
 
 function displayJCWarning() {
   var color = [
@@ -3308,25 +1782,6 @@ function displayJCWarning() {
 
 };
 
-
-function shuffle(array) {
-  let currentIndex = array.length, randomIndex;
-
-  // While there remain elements to shuffle.
-  while (currentIndex != 0) {
-
-    // Pick a remaining element.
-    randomIndex = Math.floor(Math.random() * currentIndex);
-    currentIndex--;
-
-    // And swap it with the current element.
-    [array[currentIndex], array[randomIndex]] = [
-      array[randomIndex], array[currentIndex]];
-  }
-
-  return array;
-}
-
 function warningResponse(elt, good) {
   JCwarningAlreadyDisplayed = true;
   $(elt).dialog("close");
@@ -3338,6 +1793,125 @@ function warningResponse(elt, good) {
 }
 
 
+/**
+ * *************
+ * ASSISTANT BUTTON
+ * *************
+ */
+
+$(".btnAssistant").click(function (event) {
+  var id = $(this).closest('.eqLogicDisplayCard').data('eqlogic_id');
+  openAssistantWidgetModal(id, event);
+});
+
+function openAssistantWidgetModal(id, event) {
+  if (event !== undefined) event.stopPropagation();
+  $('#md_modal').dialog({ title: "{{Configuration de l'équipement}}" });
+  $('#md_modal').load('index.php?v=d&plugin=JeedomConnect&modal=assistant.JeedomConnect&eqLogicId=' + id).dialog('open');
+}
+
+
+$(".btnNotification").click(function (event) {
+  var id = $(this).closest('.eqLogicDisplayCard').data('eqlogic_id');
+  openAssistantNotificationModal(id, event);
+});
+
+function openAssistantNotificationModal(id, event) {
+  if (event !== undefined) event.stopPropagation();
+  $('#md_modal').dialog({ title: "{{Configuration des notifications}}" });
+  $('#md_modal').load('index.php?v=d&plugin=JeedomConnect&modal=notifs.JeedomConnect&eqLogicId=' + id).dialog('open');
+}
+
+// ------------- END ASSISTANT BUTTON
+
+
+/**
+ * *************
+ * COMMAND TAB
+ * *************
+ */
+
+$("#commandtab").sortable({
+  axis: "y", cursor: "move", items: ".cmd", placeholder: "ui-state-highlight", tolerance: "intersect", forcePlaceholderSize: true
+});
+
+function addCmdToTable(_cmd) {
+  if (!isset(_cmd)) {
+    var _cmd = { configuration: {} };
+  }
+  if (!isset(_cmd.configuration)) {
+    _cmd.configuration = {};
+  }
+  var tr = '<tr class="cmd" data-cmd_id="' + init(_cmd.id) + '">';
+  tr += '<td style="min-width:50px;width:70px;">';
+  tr += '<span class="cmdAttr" data-l1key="id"></span>';
+  tr += '</td>';
+  tr += '<td style="min-width:300px;width:350px;">';
+  tr += '<div class="row">';
+  tr += '<div class="col-xs-7">';
+  tr += '<input class="cmdAttr form-control input-sm" data-l1key="name" placeholder="{{Nom de la commande}}">';
+  tr += '<select class="cmdAttr form-control input-sm" data-l1key="value" style="display : none;margin-top : 5px;" title="{{Commande information liée}}">';
+  tr += '<option value="">{{Aucune}}</option>';
+  tr += '</select>';
+  tr += '</div>';
+  tr += '<div class="col-xs-5">';
+  tr += '<a class="cmdAction btn btn-default btn-sm" data-l1key="chooseIcon"><i class="fas fa-flag"></i> {{Icône}}</a>';
+  tr += '<span class="cmdAttr" data-l1key="display" data-l2key="icon" style="margin-left : 10px;"></span>';
+  tr += '</div>';
+  tr += '</div>';
+  tr += '</td>';
+  tr += '<td>';
+  tr += '<span class="type" style="display:none;" type="' + init(_cmd.type) + '">' + jeedom.cmd.availableType() + '</span>';
+  tr += '<span class="subType" subType="' + init(_cmd.subType) + '"></span>';
+  tr += '</td>';
+  tr += '<td style="min-width:120px;width:140px;">';
+  tr += '<div><label class="checkbox-inline"><input type="checkbox" class="cmdAttr checkbox-inline" data-l1key="isVisible" checked/>{{Afficher}}</label> ';
+  tr += '<label class="checkbox-inline"><input type="checkbox" class="cmdAttr checkbox-inline" data-l1key="isHistorized" checked/>{{Historiser}}</label> ';
+  tr += '<label class="checkbox-inline"><input type="checkbox" class="cmdAttr" data-l1key="display" data-l2key="invertBinary"/>{{Inverser}}</label></div>';
+  tr += '</td>';
+  tr += '<td style="min-width:180px;">';
+  tr += '<input class="cmdAttr form-control input-sm" data-l1key="configuration" data-l2key="minValue" placeholder="{{Min.}}" title="{{Min.}}" style="width:30%;display:inline-block;"/> ';
+  tr += '<input class="cmdAttr form-control input-sm" data-l1key="configuration" data-l2key="maxValue" placeholder="{{Max.}}" title="{{Max.}}" style="width:30%;display:inline-block;"/> ';
+  tr += '<input class="cmdAttr form-control input-sm" data-l1key="unite" placeholder="{{Unité}}" title="{{Unité}}" style="width:30%;display:inline-block;"/>';
+  tr += '</td>';
+  tr += '<td >';
+  tr += '<input style="min-width:30px;width:50px;" class="cmdAttr" data-l1key="order" placeholder="{{Ordre affichage}}"/> ';
+  tr += '</td>';
+  tr += '<td>';
+  if (is_numeric(_cmd.id)) {
+    tr += '<a class="btn btn-default btn-xs cmdAction" data-action="configure"><i class="fas fa-cogs"></i></a> ';
+    tr += '<a class="btn btn-default btn-xs cmdAction" data-action="test"><i class="fas fa-rss"></i> Tester</a>';
+  }
+  tr += '<i class="fas fa-minus-circle pull-right cmdAction cursor" data-action="remove" style="display:none"></i>';
+  tr += '</td>';
+  tr += '</tr>';
+
+  $('#table_cmd tbody.cmd_' + _cmd.type).append(tr);
+  var tr = $('#table_cmd tbody.cmd_' + _cmd.type + ' tr').last();
+
+  jeedom.eqLogic.builSelectCmd({
+    id: $('.eqLogicAttr[data-l1key=id]').value(),
+    filter: { type: 'info' },
+    error: function (error) {
+      $('#div_alert').showAlert({ message: error.message, level: 'danger' });
+    },
+    success: function (result) {
+      tr.find('.cmdAttr[data-l1key=value]').append(result);
+      tr.setValues(_cmd, '.cmdAttr');
+      jeedom.cmd.changeType(tr, init(_cmd.subType));
+    }
+  });
+}
+
+// ------------- END COMMAND TAB
+
+
+/**
+ * *************
+ * DOCUMENT READY
+ * *************
+ */
+
 $(document).ready(
   function () {
     if ($(".displayJCWarning").length != 0 && (typeof JCwarningAlreadyDisplayed === 'undefined')) {
@@ -3345,3 +1919,6 @@ $(document).ready(
     }
   }
 );
+
+
+updateWidgetCount();
