+++ desktop/js/configManager.js
@@ -1121,7 +1121,8 @@ function refreshAddSummaries(_options) {
 				</div></div></li>`;
 		}
 		else if (option.category == "stringList") {
-			curOption += `<div class='input-group'><select style="width:340px;" id="${option.id}-input" onchange="subtitleSelected();">`;
+			var classSub = (option.id == "subtitle") ? "subtitleSelected" : "";
+			curOption += '<div class="input-group"><select style="width:340px;" id="${option.id}-input" class="' + classSub + '">';
 			if (!required) {
 				curOption += `<option value="none">Aucun</option>`;
 			}
